@@ -1,5 +1,6 @@
 {
   "name": "flojoy-desktop",
+  "productName": "Flojoy Studio",
   "version": "0.1.0",
   "private": true,
   "homepage": "./",
@@ -33,7 +34,6 @@
     "electron-updater": "^5.3.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "express": "^4.18.2",
-    "flojoy": "github:flojoy-ai/types",
     "html-to-image": "^1.11.11",
     "http-proxy-middleware": "^2.0.6",
     "immer": "^10.0.2",
@@ -41,7 +41,7 @@
     "jotai-immer": "^0.2.0",
     "js-yaml": "^4.1.0",
     "localforage": "^1.10.0",
-    "lucide-react": "^0.265.0",
+    "lucide-react": "^0.268.0",
     "mixpanel-browser": "^2.47.0",
     "plotly.js": "file:bauhaus/plotly.js-2.23.2.tgz",
     "rc-slider": "^10.2.1",
@@ -51,13 +51,14 @@
     "react-grid-layout": "^1.3.4",
     "react-hook-form": "^7.45.4",
     "react-json-tree": "^0.18.0",
+    "react-markdown": "^8.0.7",
     "react-plotly.js": "^2.6.0",
     "react-router-dom": "^6.14.2",
     "react-select": "^5.7.4",
     "react-syntax-highlighter": "^15.5.0",
     "react-tabs": "^6.0.2",
     "react-use": "^17.4.0",
-    "reactflow": "^11.7.4",
+    "reactflow": "^11.8.1",
     "sonner": "^0.6.2",
     "tailwind-merge": "^1.14.0",
     "upath": "^2.0.1",
@@ -102,7 +103,7 @@
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/file-saver": "^2.0.5",
-    "@types/node": "^20.4.9",
+    "@types/node": "^20.5.0",
     "@types/react": "^18.2.20",
     "@types/react-dom": "^18.2.7",
     "@types/react-grid-layout": "^1.3.2",
@@ -114,14 +115,14 @@
     "@typescript-eslint/eslint-plugin": "^6.3.0",
     "@typescript-eslint/parser": "^6.3.0",
     "@vitejs/plugin-react-swc": "^3.3.2",
-    "autoprefixer": "^10.4.14",
+    "autoprefixer": "^10.4.15",
     "concurrently": "^8.2.0",
     "cypress": "^12.17.3",
     "cypress-xpath": "^2.0.1",
     "dotenv": "^16.3.1",
     "electron": "^24.7.0",
     "electron-builder": "^24.6.3",
-    "eslint": "^8.46.0",
+    "eslint": "^8.47.0",
     "eslint-plugin-react": "^7.33.1",
     "eslint-plugin-react-refresh": "^0.4.3",
     "jest": "^29.6.2",
@@ -129,7 +130,7 @@
     "nyc": "^15.1.0",
     "postcss": "^8.4.27",
     "prettier": "3.0.1",
-    "prettier-plugin-tailwindcss": "^0.4.1",
+    "prettier-plugin-tailwindcss": "^0.5.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tailwindcss": "^3.3.3",
