@@ -14,4 +14,4 @@ ___________________
 Used for CORS configuration
 """
 # TODO - get from env variables
-origins = ["http://localhost:3000"]
+origins = ["http://localhost:5391"]
