@@ -1,4 +1,5 @@
 import logging
+import os
 import sys
 
 logger = logging.getLogger("fastapi-dev")
@@ -11,7 +12,11 @@ def logger_setup(logger: logging.Logger):
     handler = logging.StreamHandler()
     handler.setLevel(log_lvl)
 
-    formatter = CustomFormatter("%(levelname)-10s%(message)s")
+    if os.environ.get("FASTAPI_LOG"):
+        formatter = CustomFormatter("%(levelname)-10s%(message)s")
+    else:
+        formatter = CustomFormatter("%(message)s")
+
     handler.setFormatter(formatter)
 
     logger.addHandler(handler)
@@ -21,10 +26,11 @@ def get_log_level():
     map_to_int = {
         "debug": logging.DEBUG,
         "info": logging.INFO,
+        "warning": logging.WARNING,
+        "error": logging.ERROR,
+        "critical": logging.CRITICAL,
     }
-    log_level = "info"
-    if "--log-level" in sys.argv:
-        log_level = sys.argv[sys.argv.index("--log-level") + 1]
+    log_level = os.environ.get("FASTAPI_LOG", "error")
 
     return map_to_int[log_level]
 
