@@ -5,7 +5,7 @@ export default defineConfig({
   projectId: "aqkk6c",
   e2e: {
     video: false,
-    baseUrl: "http://localhost:3000/",
+    baseUrl: "http://localhost:5391/",
     setupNodeEvents(on, config) {
       return require("./cypress/plugins/index.ts")(on, config);
     },
