@@ -1,18 +1,13 @@
 import uvicorn
-import sys
 import os
 
 __ignore_list = ["venv"]
 
 if __name__ == "__main__":
-    log_level = (
-        sys.argv[sys.argv.index("--log-level") + 1]
-        if "--log-level" in sys.argv
-        else "info"
-    )
-    uvicorn.run(  # type:ignore
+    log_level = os.environ.get("FASTAPI_LOG", "error")
+    uvicorn.run(
         "captain.main:app",
-        port=8000,
+        port=5392,
         log_level=log_level,
         reload=True,
         reload_excludes=[
