@@ -11,18 +11,18 @@ const Scatter3D = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M24.237 26.2724C24.237 26.7327 24.1413 27.1315 23.9499 27.4687C23.7585 27.806 23.4988 28.0817 23.1706 28.2959C22.8471 28.5101 22.4825 28.6673 22.0769 28.7676V28.8086C22.879 28.9088 23.4874 29.1549 23.9021 29.5469C24.3214 29.9388 24.531 30.4629 24.531 31.1191C24.531 31.7025 24.3874 32.2243 24.1003 32.6846C23.8178 33.1448 23.3803 33.5071 22.7878 33.7715C22.1954 34.0358 21.432 34.168 20.4978 34.168C19.9464 34.168 19.4314 34.1224 18.9529 34.0312C18.4789 33.9446 18.0323 33.8102 17.613 33.6279V31.8301C18.0414 32.0488 18.4903 32.2152 18.9597 32.3291C19.4291 32.4385 19.8666 32.4931 20.2722 32.4931C21.0287 32.4931 21.5574 32.3633 21.8581 32.1035C22.1635 31.8392 22.3161 31.47 22.3161 30.9961C22.3161 30.7181 22.2455 30.4834 22.1042 30.292C21.963 30.1006 21.7169 29.9547 21.3659 29.8545C21.0196 29.7542 20.5342 29.7041 19.9099 29.7041H19.1511V28.084H19.9236C20.5388 28.084 21.0059 28.027 21.3249 27.9131C21.6485 27.7946 21.8672 27.6351 21.9812 27.4346C22.0997 27.2295 22.1589 26.9971 22.1589 26.7373C22.1589 26.3818 22.0495 26.1038 21.8308 25.9033C21.612 25.7028 21.2475 25.6025 20.737 25.6025C20.418 25.6025 20.1264 25.6435 19.862 25.7256C19.6023 25.803 19.3676 25.8987 19.1579 26.0127C18.9483 26.1221 18.7637 26.2292 18.6042 26.334L17.6267 24.8779C18.0186 24.5954 18.4766 24.3607 19.0007 24.1738C19.5294 23.987 20.1583 23.8935 20.8874 23.8935C21.9174 23.8935 22.7331 24.1009 23.3347 24.5156C23.9363 24.9303 24.237 25.5159 24.237 26.2724ZM35.6213 28.9385C35.6213 30.0641 35.4049 31.0052 34.9719 31.7617C34.5435 32.5137 33.9215 33.081 33.1057 33.4639C32.29 33.8421 31.3079 34.0312 30.1594 34.0312H27.3293V24.0371H30.467C31.5152 24.0371 32.4244 24.2239 33.1946 24.5976C33.9648 24.9668 34.5618 25.5159 34.9856 26.2451C35.4094 26.9697 35.6213 27.8675 35.6213 28.9385ZM33.4202 28.9931C33.4202 28.2549 33.3108 27.6487 33.092 27.1748C32.8778 26.6963 32.5588 26.3431 32.135 26.1152C31.7157 25.8874 31.1962 25.7734 30.5764 25.7734H29.4485V32.2812H30.3577C31.3922 32.2812 32.1601 32.0055 32.6614 31.4541C33.1672 30.9027 33.4202 30.0823 33.4202 28.9931ZM49.1888 31.2558C49.1888 31.8483 49.0452 32.3633 48.7581 32.8008C48.471 33.2383 48.0517 33.5755 47.5003 33.8125C46.9534 34.0495 46.288 34.168 45.5042 34.168C45.1578 34.168 44.8183 34.1452 44.4856 34.0996C44.1575 34.054 43.8408 33.9879 43.5354 33.9014C43.2347 33.8102 42.9475 33.6986 42.6741 33.5664V31.5976C43.1481 31.8073 43.6403 31.9964 44.1507 32.165C44.6611 32.3336 45.167 32.418 45.6683 32.418C46.0146 32.418 46.2926 32.3724 46.5022 32.2812C46.7164 32.1901 46.8714 32.0648 46.9671 31.9053C47.0628 31.7458 47.1106 31.5635 47.1106 31.3584C47.1106 31.1077 47.0263 30.8935 46.8577 30.7158C46.6891 30.5381 46.4567 30.3717 46.1604 30.2168C45.8688 30.0618 45.5384 29.8955 45.1692 29.7178C44.9368 29.6084 44.6839 29.4762 44.4104 29.3213C44.137 29.1618 43.8772 28.9681 43.6311 28.7402C43.385 28.5124 43.1823 28.2366 43.0227 27.9131C42.8678 27.5849 42.7903 27.193 42.7903 26.7373C42.7903 26.1403 42.927 25.6299 43.2005 25.206C43.4739 24.7822 43.8636 24.4586 44.3694 24.2353C44.8798 24.0075 45.4814 23.8935 46.1741 23.8935C46.6936 23.8935 47.1881 23.9551 47.6575 24.0781C48.1315 24.1966 48.6259 24.3698 49.1409 24.5976L48.4573 26.2451C47.997 26.0583 47.5846 25.9147 47.22 25.8144C46.8554 25.7096 46.484 25.6572 46.1058 25.6572C45.8414 25.6572 45.6158 25.7005 45.429 25.7871C45.2421 25.8691 45.1009 25.9876 45.0052 26.1426C44.9095 26.293 44.8616 26.4684 44.8616 26.6689C44.8616 26.9059 44.93 27.1064 45.0667 27.2705C45.208 27.43 45.4176 27.5849 45.6956 27.7353C45.9781 27.8857 46.3291 28.0612 46.7483 28.2617C47.2587 28.5032 47.694 28.7562 48.054 29.0205C48.4186 29.2803 48.6989 29.5879 48.8948 29.9433C49.0908 30.2943 49.1888 30.7318 49.1888 31.2558ZM56.2254 25.6572C55.8198 25.6572 55.4598 25.737 55.1453 25.8965C54.8354 26.0514 54.5734 26.277 54.3592 26.5732C54.1496 26.8695 53.99 27.2272 53.8807 27.6465C53.7713 28.0657 53.7166 28.5374 53.7166 29.0615C53.7166 29.7679 53.8032 30.3717 53.9764 30.873C54.1541 31.3698 54.4275 31.7503 54.7967 32.0146C55.1658 32.2744 55.6421 32.4043 56.2254 32.4043C56.631 32.4043 57.0366 32.3587 57.4422 32.2676C57.8524 32.1764 58.2967 32.0465 58.7752 31.8779V33.6553C58.3331 33.8375 57.8979 33.9674 57.4695 34.0449C57.0412 34.1269 56.5604 34.168 56.0272 34.168C54.9972 34.168 54.1496 33.956 53.4842 33.5322C52.8234 33.1038 52.3335 32.5068 52.0145 31.7412C51.6955 30.971 51.5359 30.0732 51.5359 29.0478C51.5359 28.2913 51.6385 27.5986 51.8436 26.9697C52.0486 26.3408 52.3494 25.7962 52.7459 25.3359C53.1424 24.8756 53.6323 24.5202 54.2156 24.2695C54.799 24.0189 55.4689 23.8935 56.2254 23.8935C56.7221 23.8935 57.2189 23.9573 57.7156 24.0849C58.2169 24.208 58.6955 24.3789 59.1512 24.5976L58.4676 26.3203C58.0939 26.1426 57.7179 25.9876 57.3397 25.8555C56.9614 25.7233 56.59 25.6572 56.2254 25.6572ZM68.0062 34.0312L67.2816 31.6523H63.638L62.9134 34.0312H60.6302L64.1576 23.9961H66.7484L70.2894 34.0312H68.0062ZM66.7757 29.875L66.0511 27.5508C66.0055 27.3958 65.944 27.1976 65.8665 26.956C65.7936 26.7099 65.7184 26.4616 65.6409 26.2109C65.568 25.9557 65.5088 25.7347 65.4632 25.5478C65.4176 25.7347 65.3538 25.9671 65.2718 26.2451C65.1943 26.5185 65.1191 26.7783 65.0462 27.0244C64.9733 27.2705 64.9209 27.4459 64.889 27.5508L64.1712 29.875H66.7757ZM76.3895 34.0312H74.2704V25.8008H71.5565V24.0371H79.1034V25.8008H76.3895V34.0312ZM85.4769 34.0312H83.3578V25.8008H80.6439V24.0371H88.1908V25.8008H85.4769V34.0312ZM96.4647 34.0312H90.7089V24.0371H96.4647V25.7734H92.828V27.9678H96.2118V29.7041H92.828V32.2812H96.4647V34.0312ZM102.449 24.0371C103.356 24.0371 104.103 24.1465 104.691 24.3652C105.283 24.584 105.723 24.9144 106.01 25.3564C106.297 25.7985 106.441 26.3568 106.441 27.0312C106.441 27.487 106.354 27.8857 106.181 28.2275C106.008 28.5693 105.78 28.8587 105.497 29.0957C105.215 29.3327 104.91 29.5264 104.581 29.6767L107.521 34.0312H105.169L102.784 30.1963H101.656V34.0312H99.5365V24.0371H102.449ZM102.298 25.7734H101.656V28.4736H102.339C103.041 28.4736 103.542 28.3574 103.843 28.125C104.149 27.888 104.301 27.5417 104.301 27.0859C104.301 26.612 104.137 26.2747 103.809 26.0742C103.485 25.8737 102.982 25.7734 102.298 25.7734Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <g clipPath="url(#clip0_5_166)">
         <path
           opacity="0.9"
           d="M47.4886 85L12 106.353H83.4382L113 85H47.4886Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M67.8298 92.7645C69.6696 92.7645 71.161 91.2613 71.161 89.407C71.161 87.5528 69.6696 86.0496 67.8298 86.0496C65.9901 86.0496 64.4987 87.5528 64.4987 89.407C64.4987 91.2613 65.9901 92.7645 67.8298 92.7645Z"
@@ -39,99 +39,99 @@ const Scatter3D = () => {
         />
         <path
           d="M58.7139 84.8831C60.446 84.8831 61.8505 83.4676 61.8505 81.7217C61.8505 79.9758 60.446 78.5603 58.7139 78.5603C56.9817 78.5603 55.5773 79.9758 55.5773 81.7217C55.5773 83.4676 56.9817 84.8831 58.7139 84.8831Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M72.654 70.8973C74.3862 70.8973 75.7906 69.4818 75.7906 67.7359C75.7906 65.9901 74.3862 64.5746 72.654 64.5746C70.9218 64.5746 69.5174 65.9901 69.5174 67.7359C69.5174 69.4818 70.9218 70.8973 72.654 70.8973Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M72.654 77.9211C74.3862 77.9211 75.7906 76.5056 75.7906 74.7597C75.7906 73.0138 74.3862 71.5983 72.654 71.5983C70.9218 71.5983 69.5174 73.0138 69.5174 74.7597C69.5174 76.5056 70.9218 77.9211 72.654 77.9211Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M33.1366 78.3227C34.8688 78.3227 36.2732 76.9072 36.2732 75.1614C36.2732 73.4155 34.8688 72 33.1366 72C31.4044 72 30 73.4155 30 75.1614C30 76.9072 31.4044 78.3227 33.1366 78.3227Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.9"
           d="M85.704 69.1058C87.4362 69.1058 88.8406 67.6903 88.8406 65.9444C88.8406 64.1986 87.4362 62.7831 85.704 62.7831C83.9718 62.7831 82.5674 64.1986 82.5674 65.9444C82.5674 67.6903 83.9718 69.1058 85.704 69.1058Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.9"
           d="M71.2602 65.9767C72.9924 65.9767 74.3968 64.5612 74.3968 62.8153C74.3968 61.0694 72.9924 59.6539 71.2602 59.6539C69.528 59.6539 68.1236 61.0694 68.1236 62.8153C68.1236 64.5612 69.528 65.9767 71.2602 65.9767Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.75"
           d="M69.1683 61.0587C70.9005 61.0587 72.3049 59.6432 72.3049 57.8973C72.3049 56.1514 70.9005 54.736 69.1683 54.736C67.4361 54.736 66.0317 56.1514 66.0317 57.8973C66.0317 59.6432 67.4361 61.0587 69.1683 61.0587Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.75"
           d="M71.2602 57.5455C72.9924 57.5455 74.3968 56.13 74.3968 54.3841C74.3968 52.6382 72.9924 51.2227 71.2602 51.2227C69.528 51.2227 68.1236 52.6382 68.1236 54.3841C68.1236 56.13 69.528 57.5455 71.2602 57.5455Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M74.7459 54.736C76.4781 54.736 77.8825 53.3205 77.8825 51.5746C77.8825 49.8287 76.4781 48.4132 74.7459 48.4132C73.0138 48.4132 71.6093 49.8287 71.6093 51.5746C71.6093 53.3205 73.0138 54.736 74.7459 54.736Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M36.9042 77.537C38.6364 77.537 40.0408 76.1215 40.0408 74.3756C40.0408 72.6298 38.6364 71.2143 36.9042 71.2143C35.1721 71.2143 33.7677 72.6298 33.7677 74.3756C33.7677 76.1215 35.1721 77.537 36.9042 77.537Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M45.885 72.3826C47.6172 72.3826 49.0216 70.9671 49.0216 69.2213C49.0216 67.4754 47.6172 66.0599 45.885 66.0599C44.1528 66.0599 42.7484 67.4754 42.7484 69.2213C42.7484 70.9671 44.1528 72.3826 45.885 72.3826Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M48.9523 60.3576C50.6845 60.3576 52.0889 58.9421 52.0889 57.1963C52.0889 55.4504 50.6845 54.0349 48.9523 54.0349C47.2201 54.0349 45.8157 55.4504 45.8157 57.1963C45.8157 58.9421 47.2201 60.3576 48.9523 60.3576Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M55.2255 77.22C56.9577 77.22 58.3621 75.8045 58.3621 74.0587C58.3621 72.3128 56.9577 70.8973 55.2255 70.8973C53.4933 70.8973 52.0889 72.3128 52.0889 74.0587C52.0889 75.8045 53.4933 77.22 55.2255 77.22Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M58.0157 74.4079C59.7478 74.4079 61.1522 72.9924 61.1522 71.2465C61.1522 69.5006 59.7478 68.0851 58.0157 68.0851C56.2835 68.0851 54.8791 69.5006 54.8791 71.2465C54.8791 72.9924 56.2835 74.4079 58.0157 74.4079Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M62.1969 74.4079C63.9291 74.4079 65.3335 72.9924 65.3335 71.2465C65.3335 69.5006 63.9291 68.0851 62.1969 68.0851C60.4647 68.0851 59.0603 69.5006 59.0603 71.2465C59.0603 72.9924 60.4647 74.4079 62.1969 74.4079Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M58.0157 54.736C59.7478 54.736 61.1522 53.3205 61.1522 51.5746C61.1522 49.8287 59.7478 48.4132 58.0157 48.4132C56.2835 48.4132 54.8791 49.8287 54.8791 51.5746C54.8791 53.3205 56.2835 54.736 58.0157 54.736Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M55.7292 85.2645C57.4613 85.2645 58.8658 83.849 58.8658 82.1031C58.8658 80.3572 57.4613 78.9417 55.7292 78.9417C53.997 78.9417 52.5926 80.3572 52.5926 82.1031C52.5926 83.849 53.997 85.2645 55.7292 85.2645Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M34.0368 61.107C35.769 61.107 37.1734 59.6915 37.1734 57.9457C37.1734 56.1998 35.769 54.7843 34.0368 54.7843C32.3046 54.7843 30.9002 56.1998 30.9002 57.9457C30.9002 59.6915 32.3046 61.107 34.0368 61.107Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M51.4946 57.9752C53.2268 57.9752 54.6312 56.5597 54.6312 54.8138C54.6312 53.068 53.2268 51.6525 51.4946 51.6525C49.7624 51.6525 48.358 53.068 48.358 54.8138C48.358 56.5597 49.7624 57.9752 51.4946 57.9752Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M33.4185 57.8624C35.1507 57.8624 36.5551 56.4469 36.5551 54.701C36.5551 52.9552 35.1507 51.5397 33.4185 51.5397C31.6864 51.5397 30.282 52.9552 30.282 54.701C30.282 56.4469 31.6864 57.8624 33.4185 57.8624Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
@@ -141,52 +141,52 @@ const Scatter3D = () => {
         <path
           opacity="0.59"
           d="M92.273 72.3504C94.0052 72.3504 95.4096 70.9349 95.4096 69.189C95.4096 67.4432 94.0052 66.0277 92.273 66.0277C90.5408 66.0277 89.1364 67.4432 89.1364 69.189C89.1364 70.9349 90.5408 72.3504 92.273 72.3504Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M101.336 66.7287C103.069 66.7287 104.473 65.3132 104.473 63.5674C104.473 61.8215 103.069 60.406 101.336 60.406C99.6042 60.406 98.1998 61.8215 98.1998 63.5674C98.1998 65.3132 99.6042 66.7287 101.336 66.7287Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M78.8232 59.3236C80.5554 59.3236 81.9598 57.9081 81.9598 56.1622C81.9598 54.4163 80.5554 53.0008 78.8232 53.0008C77.0911 53.0008 75.6866 54.4163 75.6866 56.1622C75.6866 57.9081 77.0911 59.3236 78.8232 59.3236Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M91.9799 75.4285C93.7121 75.4285 95.1165 74.013 95.1165 72.2672C95.1165 70.5213 93.7121 69.1058 91.9799 69.1058C90.2477 69.1058 88.8433 70.5213 88.8433 72.2672C88.8433 74.013 90.2477 75.4285 91.9799 75.4285Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M76.7339 69.8632C78.4661 69.8632 79.8705 68.4477 79.8705 66.7019C79.8705 64.956 78.4661 63.5405 76.7339 63.5405C75.0018 63.5405 73.5974 64.956 73.5974 66.7019C73.5974 68.4477 75.0018 69.8632 76.7339 69.8632Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M68.089 81.069C69.8212 81.069 71.2256 79.6535 71.2256 77.9076C71.2256 76.1618 69.8212 74.7463 68.089 74.7463C66.3568 74.7463 64.9524 76.1618 64.9524 77.9076C64.9524 79.6535 66.3568 81.069 68.089 81.069Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M70.8765 83.8812C72.6087 83.8812 74.0131 82.4657 74.0131 80.7198C74.0131 78.974 72.6087 77.5585 70.8765 77.5585C69.1443 77.5585 67.7399 78.974 67.7399 80.7198C67.7399 82.4657 69.1443 83.8812 70.8765 83.8812Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M86.4102 80.7547C88.1424 80.7547 89.5468 79.3393 89.5468 77.5934C89.5468 75.8475 88.1424 74.432 86.4102 74.432C84.678 74.432 83.2736 75.8475 83.2736 77.5934C83.2736 79.3393 84.678 80.7547 86.4102 80.7547Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M89.1977 77.9426C90.9299 77.9426 92.3343 76.5271 92.3343 74.7812C92.3343 73.0353 90.9299 71.6198 89.1977 71.6198C87.4655 71.6198 86.0611 73.0353 86.0611 74.7812C86.0611 76.5271 87.4655 77.9426 89.1977 77.9426Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M93.3816 77.9426C95.1138 77.9426 96.5182 76.5271 96.5182 74.7812C96.5182 73.0353 95.1138 71.6198 93.3816 71.6198C91.6495 71.6198 90.245 73.0353 90.245 74.7812C90.245 76.5271 91.6495 77.9426 93.3816 77.9426Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
