@@ -11,30 +11,30 @@ const Surface3D = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <g clipPath="url(#clip0_5_114)">
         <path
           opacity="0.64"
           d="M50.0463 60.4356L52.7161 58.4509L90.2222 65.3223L72.8807 79.4693L40.6173 70.9811L50.0463 60.4356Z"
-          className="fill-accent1"
+          className="fill-accent2"
           fillOpacity="0.9"
         />
         <path
           opacity="0.6"
           d="M45.4931 85.5686C45.4931 85.5686 50.6996 85.9405 55.9424 93.6202C61.1852 101.3 67.2346 113.426 74.8971 107.363C82.5597 101.3 88.351 86.1547 92.7146 83.2202L45.4971 85.5727L45.4931 85.5686Z"
-          className="fill-accent1"
+          className="fill-accent2"
           fillOpacity="0.8"
         />
         <path
           d="M14 78.661C14 78.661 17.6982 82.3877 31.9788 55.6661C46.2593 28.9446 51.9095 88.3617 70.4609 70.1728L81.3498 52.3881C81.3498 52.3881 87.8025 40.2621 96.6749 52.3881C105.547 64.514 105.951 69.7686 112 70.1728L78.5267 92.8079L14 78.661Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
       </g>
       <path
         d="M23.2185 26.2724C23.2185 26.8968 23.0294 27.4277 22.6511 27.8652C22.2728 28.3027 21.7419 28.6035 21.0583 28.7676V28.8086C21.865 28.9088 22.4756 29.1549 22.8904 29.5469C23.3051 29.9342 23.5124 30.4583 23.5124 31.1191C23.5124 32.0807 23.1638 32.8304 22.4665 33.3681C21.7693 33.9014 20.7735 34.168 19.4792 34.168C18.3946 34.168 17.433 33.9879 16.5945 33.6279V31.8301C16.9818 32.026 17.4079 32.1855 17.8728 32.3086C18.3376 32.4316 18.7979 32.4931 19.2536 32.4931C19.9509 32.4931 20.4659 32.3747 20.7986 32.1377C21.1312 31.9007 21.2976 31.5202 21.2976 30.9961C21.2976 30.5267 21.1062 30.194 20.7234 29.998C20.3406 29.8021 19.7299 29.7041 18.8913 29.7041H18.1325V28.084H18.905C19.6797 28.084 20.2449 27.9837 20.6003 27.7832C20.9603 27.5781 21.1404 27.2295 21.1404 26.7373C21.1404 25.9808 20.6664 25.6025 19.7185 25.6025C19.3904 25.6025 19.0554 25.6572 18.7136 25.7666C18.3764 25.876 18.0004 26.0651 17.5857 26.334L16.6081 24.8779C17.5196 24.2217 18.6065 23.8935 19.8689 23.8935C20.9034 23.8935 21.7191 24.1032 22.3161 24.5224C22.9177 24.9417 23.2185 25.525 23.2185 26.2724ZM34.5891 28.9385C34.5891 30.5836 34.1197 31.8437 33.1809 32.7187C32.2467 33.5937 30.8954 34.0312 29.1272 34.0312H26.2971V24.0371H29.4348C31.0663 24.0371 32.3333 24.4678 33.2356 25.3291C34.1379 26.1904 34.5891 27.3935 34.5891 28.9385ZM32.3879 28.9931C32.3879 26.8467 31.44 25.7734 29.5442 25.7734H28.4163V32.2812H29.3254C31.3671 32.2812 32.3879 31.1852 32.3879 28.9931ZM48.1565 31.2558C48.1565 32.1582 47.8307 32.8691 47.179 33.3887C46.5319 33.9082 45.6295 34.168 44.472 34.168C43.4056 34.168 42.4622 33.9674 41.6419 33.5664V31.5976C42.3164 31.8984 42.886 32.1103 43.3509 32.2334C43.8203 32.3564 44.2487 32.418 44.636 32.418C45.1009 32.418 45.4563 32.3291 45.7024 32.1514C45.9531 31.9736 46.0784 31.7093 46.0784 31.3584C46.0784 31.1624 46.0237 30.9892 45.9143 30.8389C45.805 30.6839 45.6432 30.5358 45.429 30.3945C45.2194 30.2532 44.7887 30.0277 44.137 29.7178C43.5263 29.4306 43.0683 29.1549 42.763 28.8906C42.4576 28.6263 42.2138 28.3187 42.0315 27.9678C41.8492 27.6168 41.7581 27.2067 41.7581 26.7373C41.7581 25.8532 42.0566 25.1582 42.6536 24.6523C43.2552 24.1465 44.0846 23.8935 45.1419 23.8935C45.6614 23.8935 46.1559 23.9551 46.6253 24.0781C47.0992 24.2012 47.5937 24.3743 48.1087 24.5976L47.4251 26.2451C46.8919 26.0264 46.4498 25.8737 46.0989 25.7871C45.7526 25.7005 45.4108 25.6572 45.0735 25.6572C44.6725 25.6572 44.3649 25.7506 44.1507 25.9375C43.9365 26.1243 43.8294 26.3681 43.8294 26.6689C43.8294 26.8558 43.8727 27.0198 43.9593 27.1611C44.0459 27.2978 44.1826 27.4323 44.3694 27.5644C44.5608 27.692 45.0097 27.9245 45.7161 28.2617C46.6503 28.7083 47.2906 29.1572 47.637 29.6084C47.9834 30.055 48.1565 30.6042 48.1565 31.2558ZM59.0828 24.0371V30.5039C59.0828 31.2422 58.9165 31.8893 58.5838 32.4453C58.2557 33.0013 57.7794 33.4274 57.1551 33.7236C56.5307 34.0198 55.7925 34.168 54.9402 34.168C53.6551 34.168 52.657 33.8398 51.9461 33.1836C51.2352 32.5228 50.8797 31.6204 50.8797 30.4765V24.0371H52.992V30.1553C52.992 30.9254 53.1469 31.4905 53.4568 31.8506C53.7667 32.2106 54.2794 32.3906 54.9949 32.3906C55.6876 32.3906 56.1889 32.2106 56.4988 31.8506C56.8133 31.486 56.9705 30.9163 56.9705 30.1416V24.0371H59.0828ZM64.6292 28.4736H65.3128C65.9827 28.4736 66.4772 28.362 66.7962 28.1387C67.1152 27.9153 67.2747 27.5644 67.2747 27.0859C67.2747 26.612 67.1107 26.2747 66.7826 26.0742C66.459 25.8737 65.9554 25.7734 65.2718 25.7734H64.6292V28.4736ZM64.6292 30.1963V34.0312H62.5101V24.0371H65.4222C66.7803 24.0371 67.7852 24.2855 68.4368 24.7822C69.0885 25.2744 69.4144 26.0241 69.4144 27.0312C69.4144 27.6191 69.2526 28.1432 68.929 28.6035C68.6055 29.0592 68.1475 29.417 67.555 29.6767C69.0589 31.9235 70.0387 33.375 70.4945 34.0312H68.1429L65.7572 30.1963H64.6292ZM74.8172 34.0312H72.7323V24.0371H78.4608V25.7734H74.8172V28.3506H78.2079V30.0801H74.8172V34.0312ZM87.514 34.0312L86.7894 31.6523H83.1459L82.4213 34.0312H80.1381L83.6654 23.9961H86.2562L89.7972 34.0312H87.514ZM86.2836 29.875C85.6136 27.7194 85.2354 26.5003 85.1488 26.2178C85.0668 25.9352 85.0075 25.7119 84.9711 25.5478C84.8207 26.1312 84.39 27.5736 83.6791 29.875H86.2836ZM96.287 25.6572C95.4895 25.6572 94.872 25.958 94.4345 26.5596C93.997 27.1566 93.7782 27.9905 93.7782 29.0615C93.7782 31.29 94.6145 32.4043 96.287 32.4043C96.9888 32.4043 97.8388 32.2288 98.8368 31.8779V33.6553C98.0165 33.9971 97.1005 34.168 96.0888 34.168C94.635 34.168 93.523 33.7282 92.7528 32.8486C91.9826 31.9645 91.5975 30.6976 91.5975 29.0478C91.5975 28.0088 91.7867 27.0996 92.1649 26.3203C92.5432 25.5364 93.0855 24.9372 93.7919 24.5224C94.5028 24.1032 95.3345 23.8935 96.287 23.8935C97.2577 23.8935 98.233 24.1282 99.2128 24.5976L98.5292 26.3203C98.1555 26.1426 97.7795 25.9876 97.4013 25.8555C97.023 25.7233 96.6516 25.6572 96.287 25.6572ZM107.705 34.0312H101.95V24.0371H107.705V25.7734H104.069V27.9678H107.453V29.7041H104.069V32.2812H107.705V34.0312Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <defs>
         <clipPath id="clip0_5_114">
