@@ -11,7 +11,7 @@ export const ArrayView = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <rect
@@ -20,7 +20,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -29,7 +29,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -38,7 +38,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -47,7 +47,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -56,7 +56,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -65,7 +65,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -74,7 +74,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -83,7 +83,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -92,7 +92,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -101,7 +101,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -110,7 +110,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -119,7 +119,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -128,7 +128,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -137,7 +137,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -146,7 +146,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -155,7 +155,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -164,7 +164,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.3"
       />
       <rect
@@ -173,7 +173,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.3"
       />
       <rect
@@ -182,7 +182,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.3"
       />
       <rect
@@ -191,7 +191,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.3"
       />
       <rect
@@ -200,7 +200,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <rect
@@ -209,7 +209,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <rect
@@ -218,7 +218,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <rect
@@ -227,7 +227,7 @@ export const ArrayView = () => {
         width="9.5"
         height="9.5"
         rx="1"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <rect
@@ -237,7 +237,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 24 101.333)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -247,7 +247,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 24 87.6111)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -257,7 +257,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 24 73.8889)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -267,7 +267,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 24 60.1667)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -277,7 +277,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 90.5 101.333)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -287,7 +287,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 90.5 87.6111)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -297,7 +297,7 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 90.5 73.8889)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -307,12 +307,12 @@ export const ArrayView = () => {
         height="9.5"
         rx="1"
         transform="rotate(-90 90.5 60.1667)"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <path
         d="M21.8097 31L21.0851 28.6211H17.4416L16.717 31H14.4338L17.9611 20.9648H20.5519L24.0929 31H21.8097ZM20.5793 26.8438C19.9093 24.6882 19.5311 23.4691 19.4445 23.1865C19.3625 22.904 19.3032 22.6807 19.2668 22.5166C19.1164 23.0999 18.6857 24.5423 17.9748 26.8438H20.5793ZM28.4567 25.4424H29.1403C29.8102 25.4424 30.3047 25.3307 30.6237 25.1074C30.9427 24.8841 31.1022 24.5332 31.1022 24.0547C31.1022 23.5807 30.9382 23.2435 30.61 23.043C30.2865 22.8424 29.7829 22.7422 29.0993 22.7422H28.4567V25.4424ZM28.4567 27.165V31H26.3376V21.0059H29.2497C30.6078 21.0059 31.6126 21.2542 32.2643 21.751C32.916 22.2432 33.2419 22.9928 33.2419 24C33.2419 24.5879 33.0801 25.112 32.7565 25.5723C32.433 26.028 31.9749 26.3857 31.3825 26.6455C32.8864 28.8923 33.8662 30.3438 34.322 31H31.9704L29.5846 27.165H28.4567ZM38.6789 25.4424H39.3625C40.0324 25.4424 40.5269 25.3307 40.8459 25.1074C41.1649 24.8841 41.3244 24.5332 41.3244 24.0547C41.3244 23.5807 41.1604 23.2435 40.8322 23.043C40.5087 22.8424 40.0051 22.7422 39.3215 22.7422H38.6789V25.4424ZM38.6789 27.165V31H36.5598V21.0059H39.4719C40.8299 21.0059 41.8348 21.2542 42.4865 21.751C43.1382 22.2432 43.4641 22.9928 43.4641 24C43.4641 24.5879 43.3023 25.112 42.9787 25.5723C42.6551 26.028 42.1971 26.3857 41.6047 26.6455C43.1086 28.8923 44.0884 30.3438 44.5441 31H42.1926L39.8068 27.165H38.6789ZM52.9001 31L52.1755 28.6211H48.532L47.8073 31H45.5241L49.0515 20.9648H51.6423L55.1833 31H52.9001ZM51.6696 26.8438C50.9997 24.6882 50.6215 23.4691 50.5349 23.1865C50.4529 22.904 50.3936 22.6807 50.3571 22.5166C50.2068 23.0999 49.7761 24.5423 49.0652 26.8438H51.6696ZM60.5383 25.1211L62.6233 21.0059H64.9065L61.5911 27.1104V31H59.4856V27.1787L56.1702 21.0059H58.467L60.5383 25.1211ZM77.4622 21.0059H79.6018L76.2044 31H73.8938L70.5032 21.0059H72.6429L74.5227 26.9531C74.6276 27.304 74.7346 27.7142 74.844 28.1836C74.958 28.6484 75.0286 28.972 75.0559 29.1543C75.1061 28.735 75.277 28.0013 75.5686 26.9531L77.4622 21.0059ZM81.8465 31V21.0059H83.9656V31H81.8465ZM93.2171 31H87.4612V21.0059H93.2171V22.7422H89.5804V24.9365H92.9642V26.6729H89.5804V29.25H93.2171V31ZM106.023 31H103.61L102.257 25.75C102.207 25.5632 102.12 25.1781 101.997 24.5947C101.878 24.0068 101.81 23.6126 101.792 23.4121C101.764 23.6582 101.696 24.0547 101.587 24.6016C101.477 25.1439 101.393 25.5312 101.334 25.7637L99.9872 31H97.5809L95.0311 21.0059H97.1161L98.3944 26.4609C98.6177 27.4681 98.7795 28.3408 98.8797 29.0791C98.9071 28.8193 98.9686 28.4183 99.0643 27.876C99.1646 27.3291 99.258 26.9053 99.3446 26.6045L100.801 21.0059H102.804L104.26 26.6045C104.323 26.8551 104.403 27.238 104.499 27.7529C104.595 28.2679 104.667 28.71 104.718 29.0791C104.763 28.7236 104.836 28.2816 104.936 27.7529C105.037 27.2197 105.128 26.7891 105.21 26.4609L106.481 21.0059H108.566L106.023 31Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
     </svg>
   );
