@@ -11,38 +11,38 @@ const Bar = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <g clipPath="url(#clip0_5_159)">
         <path
           d="M32.4212 92.7522H22V107.919H32.4212V92.7522Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M46.5665 49H36.1453V107.869H46.5665V49Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M60.7118 80.1116H50.2906V107.86H60.7118V80.1116Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M74.8571 101.507H64.4359V107.997H74.8571V101.507Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M89.0024 92.7522H78.5812V107.966H89.0024V92.7522Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M103.148 73.6061H92.7266V107.857H103.148V73.6061Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
       </g>
       <path
         d="M48.4209 24.0371H51.5313C52.9486 24.0371 53.9763 24.2399 54.6143 24.6455C55.2569 25.0465 55.5781 25.6868 55.5781 26.5664C55.5781 27.1634 55.4369 27.6533 55.1543 28.0361C54.8763 28.4189 54.5049 28.6491 54.0401 28.7265V28.7949C54.6735 28.9362 55.1293 29.2005 55.4072 29.5879C55.6898 29.9752 55.8311 30.4902 55.8311 31.1328C55.8311 32.0443 55.5007 32.7552 54.8399 33.2656C54.1836 33.776 53.2904 34.0312 52.1602 34.0312H48.4209V24.0371ZM50.5401 27.9951H51.7705C52.3447 27.9951 52.7595 27.9062 53.0147 27.7285C53.2744 27.5508 53.4043 27.2568 53.4043 26.8467C53.4043 26.4639 53.263 26.1904 52.9805 26.0264C52.7025 25.8577 52.2604 25.7734 51.6543 25.7734H50.5401V27.9951ZM50.5401 29.6767V32.2812H51.9209C52.5043 32.2812 52.9349 32.1696 53.2129 31.9463C53.4909 31.723 53.6299 31.3812 53.6299 30.9209C53.6299 30.0915 53.0375 29.6767 51.8526 29.6767H50.5401ZM64.9253 34.0312L64.2007 31.6523H60.5572L59.8326 34.0312H57.5494L61.0767 23.9961H63.6675L67.2085 34.0312H64.9253ZM63.6949 29.875C63.0249 27.7194 62.6467 26.5003 62.5601 26.2178C62.4781 25.9352 62.4188 25.7119 62.3824 25.5478C62.232 26.1312 61.8013 27.5736 61.0904 29.875H63.6949ZM71.5723 28.4736H72.2559C72.9258 28.4736 73.4203 28.362 73.7393 28.1387C74.0583 27.9153 74.2178 27.5644 74.2178 27.0859C74.2178 26.612 74.0538 26.2747 73.7256 26.0742C73.4021 25.8737 72.8985 25.7734 72.2149 25.7734H71.5723V28.4736ZM71.5723 30.1963V34.0312H69.4532V24.0371H72.3653C73.7234 24.0371 74.7282 24.2855 75.3799 24.7822C76.0316 25.2744 76.3575 26.0241 76.3575 27.0312C76.3575 27.6191 76.1957 28.1432 75.8721 28.6035C75.5486 29.0592 75.0906 29.417 74.4981 29.6767C76.002 31.9235 76.9818 33.375 77.4376 34.0312H75.086L72.7003 30.1963H71.5723Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <defs>
         <clipPath id="clip0_5_159">
