@@ -11,16 +11,16 @@ const BigNumber = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M14.5998 24.0059H17.7101C19.1274 24.0059 20.1551 24.2087 20.7931 24.6143C21.4357 25.0153 21.757 25.6556 21.757 26.5352C21.757 27.1322 21.6157 27.6221 21.3332 28.0049C21.0552 28.3877 20.6838 28.6178 20.2189 28.6953V28.7637C20.8524 28.9049 21.3081 29.1693 21.5861 29.5566C21.8686 29.944 22.0099 30.459 22.0099 31.1016C22.0099 32.013 21.6795 32.724 21.0187 33.2344C20.3625 33.7448 19.4692 34 18.339 34H14.5998V24.0059ZM16.7189 27.9639H17.9494C18.5236 27.9639 18.9383 27.875 19.1935 27.6973C19.4533 27.5195 19.5832 27.2256 19.5832 26.8154C19.5832 26.4326 19.4419 26.1592 19.1593 25.9951C18.8813 25.8265 18.4393 25.7422 17.8332 25.7422H16.7189V27.9639ZM16.7189 29.6455V32.25H18.0998C18.6831 32.25 19.1138 32.1383 19.3918 31.915C19.6698 31.6917 19.8088 31.3499 19.8088 30.8896C19.8088 30.0602 19.2163 29.6455 18.0314 29.6455H16.7189ZM24.986 34V24.0059H27.1052V34H24.986ZM34.3947 28.3877H38.3596V33.5693C37.717 33.779 37.1109 33.9248 36.5412 34.0068C35.9761 34.0934 35.3973 34.1367 34.8049 34.1367C33.2964 34.1367 32.1434 33.6947 31.3459 32.8105C30.5529 31.9219 30.1564 30.6481 30.1564 28.9893C30.1564 27.376 30.6167 26.1182 31.5373 25.2158C32.4624 24.3135 33.743 23.8623 35.3791 23.8623C36.4045 23.8623 37.3934 24.0674 38.3459 24.4775L37.6418 26.1729C36.9126 25.8083 36.1538 25.626 35.3654 25.626C34.4494 25.626 33.7157 25.9336 33.1643 26.5488C32.6128 27.1641 32.3371 27.9912 32.3371 29.0303C32.3371 30.1149 32.5581 30.9443 33.0002 31.5186C33.4468 32.0882 34.0939 32.373 34.9416 32.373C35.3837 32.373 35.8326 32.3275 36.2883 32.2363V30.1514H34.3947V28.3877ZM55.2082 34H52.5149L48.1672 26.4395H48.1057C48.1923 27.7747 48.2356 28.7272 48.2356 29.2969V34H46.342V24.0059H49.0149L53.3557 31.4912H53.4036C53.3352 30.1924 53.301 29.2741 53.301 28.7363V24.0059H55.2082V34ZM66.8455 24.0059V30.4727C66.8455 31.2109 66.6791 31.8581 66.3464 32.4141C66.0183 32.9701 65.5421 33.3962 64.9177 33.6924C64.2934 33.9886 63.5551 34.1367 62.7029 34.1367C61.4177 34.1367 60.4197 33.8086 59.7088 33.1523C58.9978 32.4915 58.6423 31.5892 58.6423 30.4453V24.0059H60.7547V30.124C60.7547 30.8942 60.9096 31.4593 61.2195 31.8193C61.5294 32.1794 62.0421 32.3594 62.7576 32.3594C63.4503 32.3594 63.9516 32.1794 64.2615 31.8193C64.5759 31.4548 64.7332 30.8851 64.7332 30.1104V24.0059H66.8455ZM74.5042 34L72.0979 26.1592H72.0364C72.123 27.7542 72.1663 28.8184 72.1663 29.3516V34H70.2727V24.0059H73.1575L75.5227 31.6484H75.5638L78.0725 24.0059H80.9573V34H78.9817V29.2695C78.9817 29.0462 78.984 28.7887 78.9886 28.4971C78.9977 28.2054 79.0296 27.4307 79.0843 26.1729H79.0227L76.4456 34H74.5042ZM84.4598 24.0059H87.5701C88.9874 24.0059 90.0151 24.2087 90.6531 24.6143C91.2957 25.0153 91.617 25.6556 91.617 26.5352C91.617 27.1322 91.4757 27.6221 91.1932 28.0049C90.9152 28.3877 90.5437 28.6178 90.0789 28.6953V28.7637C90.7124 28.9049 91.1681 29.1693 91.4461 29.5566C91.7286 29.944 91.8699 30.459 91.8699 31.1016C91.8699 32.013 91.5395 32.724 90.8787 33.2344C90.2225 33.7448 89.3292 34 88.199 34H84.4598V24.0059ZM86.5789 27.9639H87.8094C88.3836 27.9639 88.7983 27.875 89.0535 27.6973C89.3133 27.5195 89.4432 27.2256 89.4432 26.8154C89.4432 26.4326 89.3019 26.1592 89.0193 25.9951C88.7413 25.8265 88.2993 25.7422 87.6932 25.7422H86.5789V27.9639ZM86.5789 29.6455V32.25H87.9598C88.5431 32.25 88.9738 32.1383 89.2518 31.915C89.5297 31.6917 89.6687 31.3499 89.6687 30.8896C89.6687 30.0602 89.0763 29.6455 87.8914 29.6455H86.5789ZM100.602 34H94.846V24.0059H100.602V25.7422H96.9651V27.9365H100.349V29.6729H96.9651V32.25H100.602V34ZM105.793 28.4424H106.476C107.146 28.4424 107.641 28.3307 107.96 28.1074C108.279 27.8841 108.438 27.5332 108.438 27.0547C108.438 26.5807 108.274 26.2435 107.946 26.043C107.623 25.8424 107.119 25.7422 106.435 25.7422H105.793V28.4424ZM105.793 30.165V34H103.674V24.0059H106.586C107.944 24.0059 108.949 24.2542 109.6 24.751C110.252 25.2432 110.578 25.9928 110.578 27C110.578 27.5879 110.416 28.112 110.093 28.5723C109.769 29.028 109.311 29.3857 108.719 29.6455C110.223 31.8923 111.202 33.3438 111.658 34H109.306L106.921 30.165H105.793Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M25.1641 85H19.7324V70.1289L19.7852 67.6855L19.873 65.0137C18.9707 65.916 18.3438 66.5078 17.9922 66.7891L15.0391 69.1621L12.4199 65.8926L20.6992 59.3008H25.1641V85ZM32.916 82.4863C32.916 81.502 33.1797 80.7578 33.707 80.2539C34.2344 79.75 35.002 79.498 36.0098 79.498C36.9824 79.498 37.7324 79.7559 38.2598 80.2715C38.7988 80.7871 39.0684 81.5254 39.0684 82.4863C39.0684 83.4121 38.7988 84.1445 38.2598 84.6836C37.7207 85.2109 36.9707 85.4746 36.0098 85.4746C35.0254 85.4746 34.2637 85.2168 33.7246 84.7012C33.1855 84.1738 32.916 83.4355 32.916 82.4863ZM61.1113 79.6738H58.0176V85H52.709V79.6738H41.7402V75.8945L53.0078 59.3008H58.0176V75.4551H61.1113V79.6738ZM52.709 75.4551V71.0957C52.709 70.3691 52.7383 69.3145 52.7969 67.9316C52.8555 66.5488 52.9023 65.7461 52.9375 65.5234H52.7969C52.3633 66.4844 51.8418 67.4219 51.2324 68.3359L46.5215 75.4551H52.709ZM71.9922 58.9668C74.4531 58.9668 76.4336 59.5293 77.9336 60.6543C79.4453 61.7676 80.2012 63.2734 80.2012 65.1719C80.2012 66.4844 79.8379 67.6562 79.1113 68.6875C78.3848 69.707 77.2129 70.6211 75.5957 71.4297C77.5176 72.4609 78.8945 73.5391 79.7266 74.6641C80.5703 75.7773 80.9922 77.002 80.9922 78.3379C80.9922 80.4473 80.166 82.1465 78.5137 83.4355C76.8613 84.7129 74.6875 85.3516 71.9922 85.3516C69.1797 85.3516 66.9707 84.7539 65.3652 83.5586C63.7598 82.3633 62.957 80.6699 62.957 78.4785C62.957 77.0137 63.3438 75.7129 64.1172 74.5762C64.9023 73.4395 66.1562 72.4375 67.8789 71.5703C66.4141 70.6445 65.3594 69.6543 64.7148 68.5996C64.0703 67.5449 63.748 66.3906 63.748 65.1367C63.748 63.2969 64.5098 61.8086 66.0332 60.6719C67.5566 59.5352 69.543 58.9668 71.9922 58.9668ZM67.9844 78.1621C67.9844 79.1699 68.3359 79.9551 69.0391 80.5176C69.7422 81.0801 70.7031 81.3613 71.9219 81.3613C73.2695 81.3613 74.2773 81.0742 74.9453 80.5C75.6133 79.9141 75.9473 79.1465 75.9473 78.1973C75.9473 77.4121 75.6133 76.6797 74.9453 76C74.2891 75.3086 73.2168 74.5762 71.7285 73.8027C69.2324 74.9512 67.9844 76.4043 67.9844 78.1621ZM71.957 62.9395C71.0312 62.9395 70.2812 63.1797 69.707 63.6602C69.1445 64.1289 68.8633 64.7617 68.8633 65.5586C68.8633 66.2617 69.0859 66.8945 69.5312 67.457C69.9883 68.0078 70.8086 68.5762 71.9922 69.1621C73.1406 68.623 73.9434 68.0723 74.4004 67.5098C74.8574 66.9473 75.0859 66.2969 75.0859 65.5586C75.0859 64.75 74.793 64.1113 74.207 63.6426C73.6211 63.1738 72.8711 62.9395 71.957 62.9395ZM87.7949 67C87.7949 68.4883 87.9238 69.6016 88.1816 70.3398C88.4512 71.0664 88.8789 71.4297 89.4648 71.4297C90.5898 71.4297 91.1523 69.9531 91.1523 67C91.1523 64.0703 90.5898 62.6055 89.4648 62.6055C88.8789 62.6055 88.4512 62.9688 88.1816 63.6953C87.9238 64.4102 87.7949 65.5117 87.7949 67ZM95.582 66.9648C95.582 69.6602 95.0605 71.6875 94.0176 73.0469C92.9746 74.3945 91.4453 75.0684 89.4297 75.0684C87.4961 75.0684 86.002 74.377 84.9473 72.9941C83.8926 71.5996 83.3652 69.5898 83.3652 66.9648C83.3652 61.6094 85.3867 58.9316 89.4297 58.9316C91.4102 58.9316 92.9277 59.6289 93.9824 61.0234C95.0488 62.4062 95.582 64.3867 95.582 66.9648ZM107.676 59.3008L93.4199 85H89.2012L103.457 59.3008H107.676ZM105.795 77.2656C105.795 78.7539 105.924 79.8672 106.182 80.6055C106.451 81.332 106.879 81.6953 107.465 81.6953C108.59 81.6953 109.152 80.2188 109.152 77.2656C109.152 74.3359 108.59 72.8711 107.465 72.8711C106.879 72.8711 106.451 73.2344 106.182 73.9609C105.924 74.6758 105.795 75.7773 105.795 77.2656ZM113.582 77.2305C113.582 79.9141 113.061 81.9355 112.018 83.2949C110.975 84.6426 109.445 85.3164 107.43 85.3164C105.496 85.3164 104.002 84.625 102.947 83.2422C101.893 81.8477 101.365 79.8438 101.365 77.2305C101.365 71.875 103.387 69.1973 107.43 69.1973C109.41 69.1973 110.928 69.8945 111.982 71.2891C113.049 72.6719 113.582 74.6523 113.582 77.2305Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
     </svg>
   );
