@@ -11,36 +11,36 @@ const Histogram = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
-      <path d="M32 58H22V108H32V58Z" className="fill-accent1" />
+      <path d="M32 58H22V108H32V58Z" className="fill-accent2" />
       <path
         d="M42.4212 49H32V107.869H42.4212V49Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.5"
       />
       <path
         d="M62 77H52V108H62V77Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.5"
       />
-      <path d="M52 85H42V108H52V85Z" className="fill-accent1" />
-      <path d="M72 99H62V108H72V99Z" className="fill-accent1" />
+      <path d="M52 85H42V108H52V85Z" className="fill-accent2" />
+      <path d="M72 99H62V108H72V99Z" className="fill-accent2" />
       <path
         d="M82 68H72V108H82V68Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.5"
       />
       <path
         d="M102 52H92V108H102V52Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.5"
       />
-      <path d="M92 81H82V108H92V81Z" className="fill-accent1" />
+      <path d="M92 81H82V108H92V81Z" className="fill-accent2" />
       <path
         d="M25.5405 34.0312H23.4282V29.7178H19.4702V34.0312H17.3511V24.0371H19.4702V27.9541H23.4282V24.0371H25.5405V34.0312ZM29.0498 34.0312V24.0371H31.169V34.0312H29.0498ZM40.564 31.2559C40.564 32.1582 40.2381 32.8691 39.5865 33.3887C38.9393 33.9082 38.037 34.168 36.8794 34.168C35.813 34.168 34.8697 33.9674 34.0493 33.5664V31.5977C34.7238 31.8984 35.2935 32.1104 35.7583 32.2334C36.2277 32.3564 36.6561 32.418 37.0435 32.418C37.5083 32.418 37.8638 32.3291 38.1099 32.1514C38.3605 31.9736 38.4859 31.7093 38.4859 31.3584C38.4859 31.1624 38.4312 30.9893 38.3218 30.8389C38.2124 30.6839 38.0506 30.5358 37.8365 30.3945C37.6268 30.2533 37.1962 30.0277 36.5445 29.7178C35.9338 29.4307 35.4758 29.1549 35.1704 28.8906C34.8651 28.6263 34.6213 28.3187 34.439 27.9678C34.2567 27.6169 34.1656 27.2067 34.1656 26.7373C34.1656 25.8532 34.4641 25.1582 35.0611 24.6523C35.6626 24.1465 36.4921 23.8936 37.5493 23.8936C38.0689 23.8936 38.5633 23.9551 39.0327 24.0781C39.5067 24.2012 40.0012 24.3743 40.5161 24.5977L39.8325 26.2451C39.2993 26.0264 38.8573 25.8737 38.5064 25.7871C38.16 25.7005 37.8182 25.6572 37.481 25.6572C37.0799 25.6572 36.7723 25.7507 36.5581 25.9375C36.3439 26.1243 36.2368 26.3682 36.2368 26.6689C36.2368 26.8558 36.2801 27.0199 36.3667 27.1611C36.4533 27.2979 36.59 27.4323 36.7769 27.5645C36.9683 27.6921 37.4172 27.9245 38.1236 28.2617C39.0578 28.7083 39.6981 29.1572 40.0445 29.6084C40.3908 30.055 40.564 30.6042 40.564 31.2559ZM47.211 34.0312H45.0918V25.8008H42.378V24.0371H49.9249V25.8008H47.211V34.0312ZM61.5142 29.0205C61.5142 30.6748 61.1041 31.9463 60.2838 32.835C59.4634 33.7236 58.2877 34.168 56.7564 34.168C55.2252 34.168 54.0494 33.7236 53.2291 32.835C52.4088 31.9463 51.9986 30.6702 51.9986 29.0068C51.9986 27.3434 52.4088 26.0742 53.2291 25.1992C54.0539 24.3197 55.2343 23.8799 56.7701 23.8799C58.3059 23.8799 59.4794 24.3219 60.2906 25.2061C61.1063 26.0902 61.5142 27.3617 61.5142 29.0205ZM54.2203 29.0205C54.2203 30.137 54.4322 30.9779 54.856 31.543C55.2799 32.1081 55.9133 32.3906 56.7564 32.3906C58.4472 32.3906 59.2925 31.2673 59.2925 29.0205C59.2925 26.7692 58.4517 25.6436 56.7701 25.6436C55.927 25.6436 55.2912 25.9284 54.8629 26.498C54.4345 27.0632 54.2203 27.904 54.2203 29.0205ZM68.3595 28.4189H72.3243V33.6006C71.6817 33.8102 71.0756 33.9561 70.5059 34.0381C69.9408 34.1247 69.3621 34.168 68.7696 34.168C67.2612 34.168 66.1082 33.7259 65.3106 32.8418C64.5177 31.9531 64.1212 30.6794 64.1212 29.0205C64.1212 27.4072 64.5815 26.1494 65.502 25.2471C66.4272 24.3447 67.7078 23.8936 69.3438 23.8936C70.3692 23.8936 71.3582 24.0986 72.3106 24.5088L71.6065 26.2041C70.8774 25.8395 70.1186 25.6572 69.3302 25.6572C68.4142 25.6572 67.6804 25.9648 67.129 26.5801C66.5776 27.1953 66.3018 28.0225 66.3018 29.0615C66.3018 30.1462 66.5229 30.9756 66.9649 31.5498C67.4115 32.1195 68.0587 32.4043 68.9063 32.4043C69.3484 32.4043 69.7973 32.3587 70.253 32.2676V30.1826H68.3595V28.4189ZM77.8092 28.4736H78.4928C79.1627 28.4736 79.6572 28.362 79.9762 28.1387C80.2952 27.9154 80.4547 27.5645 80.4547 27.0859C80.4547 26.612 80.2906 26.2747 79.9625 26.0742C79.6389 25.8737 79.1354 25.7734 78.4518 25.7734H77.8092V28.4736ZM77.8092 30.1963V34.0312H75.69V24.0371H78.6022C79.9602 24.0371 80.9651 24.2855 81.6168 24.7822C82.2685 25.2744 82.5943 26.0241 82.5943 27.0312C82.5943 27.6191 82.4326 28.1432 82.109 28.6035C81.7854 29.0592 81.3274 29.417 80.735 29.6768C82.2389 31.9235 83.2187 33.375 83.6744 34.0312H81.3229L78.9371 30.1963H77.8092ZM92.0304 34.0312L91.3058 31.6523H87.6622L86.9376 34.0312H84.6544L88.1818 23.9961H90.7726L94.3136 34.0312H92.0304ZM90.7999 29.875C90.13 27.7194 89.7518 26.5003 89.6652 26.2178C89.5831 25.9352 89.5239 25.7119 89.4874 25.5479C89.337 26.1312 88.9064 27.5736 88.1954 29.875H90.7999ZM100.79 34.0312L98.3834 26.1904H98.3219C98.4085 27.7855 98.4518 28.8496 98.4518 29.3828V34.0312H96.5582V24.0371H99.443L101.808 31.6797H101.849L104.358 24.0371H107.243V34.0312H105.267V29.3008C105.267 29.0775 105.27 28.82 105.274 28.5283C105.283 28.2367 105.315 27.4619 105.37 26.2041H105.308L102.731 34.0312H100.79Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
     </svg>
   );
