@@ -1,6 +1,6 @@
 import { memo } from "react";
 
-export const Image = () => {
+const Image = () => {
   return (
     <svg
       width="225"
@@ -11,33 +11,33 @@ export const Image = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M30.6336 34.0312V24.0371H32.7527V34.0312H30.6336ZM40.4798 34.0312L38.0735 26.1904H38.012C38.0986 27.7855 38.1419 28.8496 38.1419 29.3828V34.0312H36.2483V24.0371H39.1331L41.4983 31.6797H41.5393L44.0481 24.0371H46.9329V34.0312H44.9573V29.3008C44.9573 29.0775 44.9596 28.82 44.9641 28.5283C44.9733 28.2366 45.0052 27.4619 45.0599 26.2041H44.9983L42.4212 34.0312H40.4798ZM56.3348 31.2558C56.3348 32.1582 56.0089 32.8691 55.3572 33.3887C54.7101 33.9082 53.8078 34.168 52.6502 34.168C51.5838 34.168 50.6404 33.9674 49.8201 33.5664V31.5976C50.4946 31.8984 51.0643 32.1103 51.5291 32.2334C51.9985 32.3564 52.4269 32.418 52.8143 32.418C53.2791 32.418 53.6346 32.3291 53.8807 32.1514C54.1313 31.9736 54.2566 31.7093 54.2566 31.3584C54.2566 31.1624 54.202 30.9892 54.0926 30.8389C53.9832 30.6839 53.8214 30.5358 53.6072 30.3945C53.3976 30.2532 52.9669 30.0277 52.3152 29.7178C51.7046 29.4306 51.2466 29.1549 50.9412 28.8906C50.6359 28.6263 50.3921 28.3187 50.2098 27.9678C50.0275 27.6168 49.9363 27.2067 49.9363 26.7373C49.9363 25.8532 50.2348 25.1582 50.8318 24.6523C51.4334 24.1465 52.2628 23.8935 53.3201 23.8935C53.8397 23.8935 54.3341 23.9551 54.8035 24.0781C55.2775 24.2012 55.7719 24.3743 56.2869 24.5976L55.6033 26.2451C55.0701 26.0264 54.6281 25.8737 54.2772 25.7871C53.9308 25.7005 53.589 25.6572 53.2518 25.6572C52.8507 25.6572 52.5431 25.7506 52.3289 25.9375C52.1147 26.1243 52.0076 26.3681 52.0076 26.6689C52.0076 26.8558 52.0509 27.0198 52.1375 27.1611C52.2241 27.2978 52.3608 27.4323 52.5477 27.5644C52.7391 27.692 53.188 27.9245 53.8943 28.2617C54.8286 28.7083 55.4689 29.1572 55.8152 29.6084C56.1616 30.055 56.3348 30.6042 56.3348 31.2558ZM67.3158 34.0312H65.2034V29.7178H61.2454V34.0312H59.1263V24.0371H61.2454V27.9541H65.2034V24.0371H67.3158V34.0312ZM79.8963 29.0205C79.8963 30.6748 79.4862 31.9463 78.6659 32.8349C77.8456 33.7236 76.6698 34.168 75.1385 34.168C73.6073 34.168 72.4315 33.7236 71.6112 32.8349C70.7909 31.9463 70.3807 30.6702 70.3807 29.0068C70.3807 27.3434 70.7909 26.0742 71.6112 25.1992C72.436 24.3196 73.6164 23.8799 75.1522 23.8799C76.688 23.8799 77.8615 24.3219 78.6727 25.206C79.4885 26.0902 79.8963 27.3616 79.8963 29.0205ZM72.6024 29.0205C72.6024 30.137 72.8143 30.9778 73.2381 31.543C73.662 32.1081 74.2954 32.3906 75.1385 32.3906C76.8293 32.3906 77.6747 31.2672 77.6747 29.0205C77.6747 26.7692 76.8338 25.6435 75.1522 25.6435C74.3091 25.6435 73.6734 25.9284 73.245 26.498C72.8166 27.0631 72.6024 27.904 72.6024 29.0205ZM92.682 34.0312H90.2689L88.9154 28.7812C88.8653 28.5944 88.7787 28.2093 88.6556 27.626C88.5372 27.0381 88.4688 26.6439 88.4506 26.4433C88.4232 26.6894 88.3549 27.0859 88.2455 27.6328C88.1361 28.1751 88.0518 28.5625 87.9926 28.7949L86.6459 34.0312H84.2396L81.6898 24.0371H83.7748L85.0531 29.4922C85.2764 30.4993 85.4382 31.3721 85.5385 32.1103C85.5658 31.8506 85.6273 31.4495 85.723 30.9072C85.8233 30.3603 85.9167 29.9365 86.0033 29.6357L87.4594 24.0371H89.4623L90.9183 29.6357C90.9821 29.8864 91.0619 30.2692 91.1576 30.7842C91.2533 31.2991 91.3262 31.7412 91.3763 32.1103C91.4219 31.7549 91.4948 31.3128 91.5951 30.7842C91.6954 30.251 91.7865 29.8203 91.8685 29.4922L93.14 24.0371H95.225L92.682 34.0312Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <rect
         x="29"
         y="42"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <rect
         x="51.3333"
         y="86.6667"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <rect
         x="51.3333"
         y="64.3333"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.5"
       />
       <rect
@@ -45,7 +45,7 @@ export const Image = () => {
         y="64.3333"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -53,7 +53,7 @@ export const Image = () => {
         y="86.6667"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -61,7 +61,7 @@ export const Image = () => {
         y="42"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -69,7 +69,7 @@ export const Image = () => {
         y="86.6667"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -77,7 +77,7 @@ export const Image = () => {
         y="64.3333"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.7"
       />
       <rect
@@ -85,7 +85,7 @@ export const Image = () => {
         y="42"
         width="22.3333"
         height="22.3333"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
     </svg>
