@@ -11,19 +11,19 @@ const LineChart = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <g clipPath="url(#clip0_5_188)">
         <path
           d="M111.767 66.8638L97.2327 76.7416L83.2441 66.4731L69.2556 80.1248L55.267 65.5529L41.2784 95.6368L27.2898 89.718L13.1364 98.7087"
-          className="stroke-accent1"
-          stroke-miterlimit="10"
+          className="stroke-accent2"
+          strokeMiterlimit="10"
         />
         <path
           d="M111.767 57.5685L97.2327 82.4353L83.2441 75.0733L69.2556 84.587L55.267 86.123L41.2784 70.1608L27.2898 76.6026L13.3979 54.192"
           stroke="white"
-          stroke-miterlimit="10"
+          strokeMiterlimit="10"
         />
         <path
           d="M28.2334 89.718C28.2334 90.3204 27.8128 90.817 27.2898 90.817C26.7669 90.817 26.3463 90.3271 26.3463 89.718C26.3463 89.1089 26.7669 88.619 27.2898 88.619C27.8128 88.619 28.2334 89.1089 28.2334 89.718Z"
@@ -76,7 +76,7 @@ const LineChart = () => {
       </g>
       <path
         d="M46.1946 34.0312V24.0371H48.3138V32.2812H52.3675V34.0312H46.1946ZM55.0906 34.0312V24.0371H57.2098V34.0312H55.0906ZM69.5716 34.0312H66.8783L62.5306 26.4707H62.4691C62.5557 27.806 62.599 28.7584 62.599 29.3281V34.0312H60.7054V24.0371H63.3783L67.7191 31.5224H67.7669C67.6986 30.2236 67.6644 29.3053 67.6644 28.7676V24.0371H69.5716V34.0312ZM78.8299 34.0312H73.0741V24.0371H78.8299V25.7734H75.1932V27.9678H78.577V29.7041H75.1932V32.2812H78.8299V34.0312Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <defs>
         <clipPath id="clip0_5_188">
