@@ -11,27 +11,27 @@ const MatrixView = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M19 43C19 41.8954 19.8954 41 21 41H44C45.1046 41 46 41.8954 46 43V53C46 54.1046 45.1046 55 44 55H21C19.8954 55 19 54.1046 19 53V43Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <path
         d="M81 62C81 60.8954 81.8954 60 83 60H106C107.105 60 108 60.8954 108 62V72C108 73.1046 107.105 74 106 74H83C81.8954 74 81 73.1046 81 72V62Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <path
         d="M81 43C81 41.8954 81.8954 41 83 41H106C107.105 41 108 41.8954 108 43V53C108 54.1046 107.105 55 106 55H83C81.8954 55 81 54.1046 81 53V43Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <path
         d="M50 88C50 86.8954 50.8954 86 52 86H75C76.1046 86 77 86.8954 77 88V98C77 99.1046 76.1046 100 75 100H52C50.8954 100 50 99.1046 50 98V88Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -40,7 +40,7 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.9"
       />
       <rect
@@ -49,7 +49,7 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <rect
@@ -58,7 +58,7 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.2"
       />
       <rect
@@ -67,7 +67,7 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -76,7 +76,7 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -85,7 +85,7 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -94,12 +94,12 @@ const MatrixView = () => {
         width="27"
         height="14"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <path
         d="M23.6672 33L21.6047 26.2793H21.552C21.6262 27.6465 21.6633 28.5586 21.6633 29.0156V33H20.0402V24.4336H22.5129L24.5402 30.9844H24.5754L26.7258 24.4336H29.1984V33H27.5051V28.9453C27.5051 28.7539 27.507 28.5332 27.5109 28.2832C27.5187 28.0332 27.5461 27.3691 27.593 26.291H27.5402L25.3312 33H23.6672ZM37.4447 33L36.8236 30.9609H33.7005L33.0795 33H31.1224L34.1459 24.3984H36.3666L39.4017 33H37.4447ZM36.39 29.4375C35.8158 27.5898 35.4916 26.5449 35.4173 26.3027C35.347 26.0605 35.2962 25.8691 35.265 25.7285C35.1361 26.2285 34.767 27.4648 34.1576 29.4375H36.39ZM44.6304 33H42.814V25.9453H40.4878V24.4336H46.9566V25.9453H44.6304V33ZM50.9313 28.2363H51.5173C52.0915 28.2363 52.5153 28.1406 52.7887 27.9492C53.0622 27.7578 53.1989 27.457 53.1989 27.0469C53.1989 26.6406 53.0583 26.3516 52.777 26.1797C52.4997 26.0078 52.068 25.9219 51.4821 25.9219H50.9313V28.2363ZM50.9313 29.7129V33H49.1149V24.4336H51.611C52.7751 24.4336 53.6364 24.6465 54.195 25.0723C54.7536 25.4941 55.0329 26.1367 55.0329 27C55.0329 27.5039 54.8942 27.9531 54.6169 28.3477C54.3395 28.7383 53.947 29.0449 53.4391 29.2676C54.7282 31.1934 55.568 32.4375 55.9587 33H53.943L51.8981 29.7129H50.9313ZM57.8768 33V24.4336H59.6932V33H57.8768ZM69.6152 33H67.541L65.5488 29.7598L63.5566 33H61.6113L64.4531 28.582L61.793 24.4336H63.7969L65.6426 27.5156L67.4531 24.4336H69.4102L66.7207 28.6816L69.6152 33ZM80.3773 24.4336H82.2113L79.2991 33H77.3187L74.4124 24.4336H76.2464L77.8577 29.5312C77.9476 29.832 78.0394 30.1836 78.1331 30.5859C78.2308 30.9844 78.2913 31.2617 78.3148 31.418C78.3577 31.0586 78.5042 30.4297 78.7542 29.5312L80.3773 24.4336ZM84.1352 33V24.4336H85.9516V33H84.1352ZM93.8815 33H88.9479V24.4336H93.8815V25.9219H90.7643V27.8027H93.6647V29.291H90.7643V31.5H93.8815V33ZM104.858 33H102.79L101.63 28.5C101.587 28.3398 101.512 28.0098 101.407 27.5098C101.305 27.0059 101.247 26.668 101.231 26.4961C101.208 26.707 101.149 27.0469 101.055 27.5156C100.962 27.9805 100.889 28.3125 100.839 28.5117L99.6844 33H97.6219L95.4363 24.4336H97.2234L98.3191 29.1094C98.5105 29.9727 98.6492 30.7207 98.7352 31.3535C98.7586 31.1309 98.8113 30.7871 98.8934 30.3223C98.9793 29.8535 99.0594 29.4902 99.1336 29.2324L100.382 24.4336H102.098L103.346 29.2324C103.401 29.4473 103.47 29.7754 103.552 30.2168C103.634 30.6582 103.696 31.0371 103.739 31.3535C103.778 31.0488 103.841 30.6699 103.927 30.2168C104.012 29.7598 104.091 29.3906 104.161 29.1094L105.251 24.4336H107.038L104.858 33Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
     </svg>
   );
