@@ -11,25 +11,25 @@ const ProphetComponents = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M11 56.8235L24.5726 66.9706L39.7903 48.7059L48.4274 60.2059L52.9516 54.1176L61.5887 66.9706L73.1048 46L80.0968 63.5882L85.4435 54.1176L91.6129 63.5882L101.484 51.4118L113 69"
-        className="stroke-accent1"
+        className="stroke-accent2"
       />
       <path
         d="M11 99.1326C13.2024 89.1555 18.1854 72.8852 20.498 87.6206C23.3887 106.04 27.5182 100.284 27.9312 89.923C28.3441 79.5622 31.6478 81.8646 31.2348 85.3182C30.8219 88.7718 34.5385 91.8417 34.5385 83.3995C34.5385 74.9574 40.3198 64.9804 44.0364 83.3995C47.753 101.819 52.7085 85.3182 53.9474 87.6206C55.1862 89.923 58.0769 105.656 63.0324 94.1441C67.9879 82.6321 70.4656 82.6321 75.0081 89.923C79.5506 97.2139 84.0931 100.668 87.8097 94.1441C91.5263 87.6206 95.6559 81.8646 101.024 93.3766C105.319 102.586 110.798 101.563 113 99.9001"
-        className="stroke-accent1"
+        className="stroke-accent2"
       />
       <path
         d="M105.463 73.0729L101.305 51L115.307 61.0591L105.463 73.0729Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.3"
       />
       <path
         d="M12.6892 28.9766H13.0877C13.4601 28.9766 13.7387 28.9036 13.9236 28.7578C14.1085 28.6094 14.2009 28.3945 14.2009 28.1133C14.2009 27.8294 14.1228 27.6198 13.9666 27.4844C13.8129 27.349 13.5707 27.2812 13.24 27.2812H12.6892V28.9766ZM15.4236 28.0703C15.4236 28.6849 15.2309 29.1549 14.8455 29.4805C14.4627 29.806 13.9171 29.9688 13.2088 29.9688H12.6892V32H11.4783V26.2891H13.3025C13.9952 26.2891 14.5213 26.4388 14.8806 26.7383C15.2426 27.0352 15.4236 27.4792 15.4236 28.0703ZM18.2727 28.8242H18.6633C19.0461 28.8242 19.3286 28.7604 19.5109 28.6328C19.6932 28.5052 19.7844 28.3047 19.7844 28.0312C19.7844 27.7604 19.6906 27.5677 19.5031 27.4531C19.3182 27.3385 19.0305 27.2812 18.6398 27.2812H18.2727V28.8242ZM18.2727 29.8086V32H17.0617V26.2891H18.7258C19.5018 26.2891 20.076 26.431 20.4484 26.7148C20.8208 26.9961 21.007 27.4245 21.007 28C21.007 28.3359 20.9146 28.6354 20.7297 28.8984C20.5448 29.1589 20.2831 29.3633 19.9445 29.5117C20.8039 30.7956 21.3638 31.625 21.6242 32H20.2805L18.9172 29.8086H18.2727ZM28.0866 29.1367C28.0866 30.082 27.8522 30.8086 27.3834 31.3164C26.9147 31.8242 26.2428 32.0781 25.3678 32.0781C24.4928 32.0781 23.8209 31.8242 23.3522 31.3164C22.8834 30.8086 22.6491 30.0794 22.6491 29.1289C22.6491 28.1784 22.8834 27.4531 23.3522 26.9531C23.8235 26.4505 24.498 26.1992 25.3756 26.1992C26.2532 26.1992 26.9238 26.4518 27.3873 26.957C27.8535 27.4622 28.0866 28.1888 28.0866 29.1367ZM23.9186 29.1367C23.9186 29.7747 24.0397 30.2552 24.2819 30.5781C24.5241 30.901 24.886 31.0625 25.3678 31.0625C26.334 31.0625 26.817 30.4206 26.817 29.1367C26.817 27.8503 26.3366 27.207 25.3756 27.207C24.8939 27.207 24.5306 27.3698 24.2858 27.6953C24.041 28.0182 23.9186 28.4987 23.9186 29.1367ZM31.0411 28.9766H31.4395C31.8119 28.9766 32.0906 28.9036 32.2755 28.7578C32.4604 28.6094 32.5528 28.3945 32.5528 28.1133C32.5528 27.8294 32.4747 27.6198 32.3184 27.4844C32.1648 27.349 31.9226 27.2812 31.5919 27.2812H31.0411V28.9766ZM33.7755 28.0703C33.7755 28.6849 33.5828 29.1549 33.1973 29.4805C32.8145 29.806 32.269 29.9688 31.5606 29.9688H31.0411V32H29.8302V26.2891H31.6544C32.3471 26.2891 32.8731 26.4388 33.2325 26.7383C33.5945 27.0352 33.7755 27.4792 33.7755 28.0703ZM40.0933 32H38.8863V29.5352H36.6245V32H35.4136V26.2891H36.6245V28.5273H38.8863V26.2891H40.0933V32ZM45.3877 32H42.0986V26.2891H45.3877V27.2812H43.3095V28.5352H45.2431V29.5273H43.3095V31H45.3877V32ZM49.3461 32H48.1352V27.2969H46.5844V26.2891H50.8969V27.2969H49.3461V32ZM57.3997 27.2148C56.944 27.2148 56.5911 27.3867 56.3411 27.7305C56.0911 28.0716 55.9661 28.5482 55.9661 29.1602C55.9661 30.4336 56.444 31.0703 57.3997 31.0703C57.8007 31.0703 58.2864 30.9701 58.8567 30.7695V31.7852C58.388 31.9805 57.8645 32.0781 57.2864 32.0781C56.4557 32.0781 55.8203 31.8268 55.3802 31.3242C54.9401 30.819 54.72 30.0951 54.72 29.1523C54.72 28.5586 54.8281 28.0391 55.0442 27.5938C55.2604 27.1458 55.5703 26.8034 55.9739 26.5664C56.3802 26.3268 56.8554 26.207 57.3997 26.207C57.9544 26.207 58.5117 26.3411 59.0716 26.6094L58.6809 27.5938C58.4674 27.4922 58.2526 27.4036 58.0364 27.3281C57.8203 27.2526 57.608 27.2148 57.3997 27.2148ZM65.8191 29.1367C65.8191 30.082 65.5847 30.8086 65.1159 31.3164C64.6472 31.8242 63.9753 32.0781 63.1003 32.0781C62.2253 32.0781 61.5534 31.8242 61.0847 31.3164C60.6159 30.8086 60.3816 30.0794 60.3816 29.1289C60.3816 28.1784 60.6159 27.4531 61.0847 26.9531C61.556 26.4505 62.2305 26.1992 63.1081 26.1992C63.9857 26.1992 64.6563 26.4518 65.1198 26.957C65.586 27.4622 65.8191 28.1888 65.8191 29.1367ZM61.6511 29.1367C61.6511 29.7747 61.7722 30.2552 62.0144 30.5781C62.2566 30.901 62.6185 31.0625 63.1003 31.0625C64.0665 31.0625 64.5495 30.4206 64.5495 29.1367C64.5495 27.8503 64.0691 27.207 63.1081 27.207C62.6264 27.207 62.2631 27.3698 62.0183 27.6953C61.7735 28.0182 61.6511 28.4987 61.6511 29.1367ZM69.9806 32L68.6056 27.5195H68.5705C68.6199 28.431 68.6447 29.0391 68.6447 29.3438V32H67.5627V26.2891H69.2111L70.5627 30.6562H70.5861L72.0197 26.2891H73.6681V32H72.5392V29.2969C72.5392 29.1693 72.5405 29.0221 72.5431 28.8555C72.5483 28.6888 72.5666 28.2461 72.5978 27.5273H72.5627L71.09 32H69.9806ZM76.8805 28.9766H77.2789C77.6513 28.9766 77.93 28.9036 78.1149 28.7578C78.2997 28.6094 78.3922 28.3945 78.3922 28.1133C78.3922 27.8294 78.3141 27.6198 78.1578 27.4844C78.0042 27.349 77.762 27.2812 77.4313 27.2812H76.8805V28.9766ZM79.6149 28.0703C79.6149 28.6849 79.4221 29.1549 79.0367 29.4805C78.6539 29.806 78.1083 29.9688 77.4 29.9688H76.8805V32H75.6695V26.2891H77.4938C78.1865 26.2891 78.7125 26.4388 79.0719 26.7383C79.4339 27.0352 79.6149 27.4792 79.6149 28.0703ZM86.4366 29.1367C86.4366 30.082 86.2022 30.8086 85.7334 31.3164C85.2647 31.8242 84.5928 32.0781 83.7178 32.0781C82.8428 32.0781 82.1709 31.8242 81.7022 31.3164C81.2334 30.8086 80.9991 30.0794 80.9991 29.1289C80.9991 28.1784 81.2334 27.4531 81.7022 26.9531C82.1735 26.4505 82.848 26.1992 83.7256 26.1992C84.6032 26.1992 85.2738 26.4518 85.7374 26.957C86.2035 27.4622 86.4366 28.1888 86.4366 29.1367ZM82.2686 29.1367C82.2686 29.7747 82.3897 30.2552 82.6319 30.5781C82.8741 30.901 83.236 31.0625 83.7178 31.0625C84.684 31.0625 85.167 30.4206 85.167 29.1367C85.167 27.8503 84.6866 27.207 83.7256 27.207C83.2439 27.207 82.8806 27.3698 82.6358 27.6953C82.391 28.0182 82.2686 28.4987 82.2686 29.1367ZM93.2466 32H91.7075L89.2231 27.6797H89.188C89.2375 28.4427 89.2622 28.987 89.2622 29.3125V32H88.1802V26.2891H89.7075L92.188 30.5664H92.2153C92.1763 29.8242 92.1567 29.2995 92.1567 28.9922V26.2891H93.2466V32ZM98.537 32H95.248V26.2891H98.537V27.2812H96.4589V28.5352H98.3925V29.5273H96.4589V31H98.537V32ZM105.359 32H103.82L101.335 27.6797H101.3C101.35 28.4427 101.374 28.987 101.374 29.3125V32H100.292V26.2891H101.82L104.3 30.5664H104.327C104.288 29.8242 104.269 29.2995 104.269 28.9922V26.2891H105.359V32ZM109.563 32H108.352V27.2969H106.802V26.2891H111.114V27.2969H109.563V32ZM115.924 30.4141C115.924 30.9297 115.738 31.3359 115.365 31.6328C114.996 31.9297 114.48 32.0781 113.819 32.0781C113.209 32.0781 112.67 31.9635 112.201 31.7344V30.6094C112.587 30.7812 112.912 30.9023 113.178 30.9727C113.446 31.043 113.691 31.0781 113.912 31.0781C114.178 31.0781 114.381 31.0273 114.522 30.9258C114.665 30.8242 114.737 30.6732 114.737 30.4727C114.737 30.3607 114.705 30.2617 114.643 30.1758C114.58 30.0872 114.488 30.0026 114.365 29.9219C114.246 29.8411 114 29.7122 113.627 29.5352C113.278 29.3711 113.017 29.2135 112.842 29.0625C112.668 28.9115 112.528 28.7357 112.424 28.5352C112.32 28.3346 112.268 28.1003 112.268 27.832C112.268 27.3268 112.438 26.9297 112.78 26.6406C113.123 26.3516 113.597 26.207 114.201 26.207C114.498 26.207 114.781 26.2422 115.049 26.3125C115.32 26.3828 115.602 26.4818 115.897 26.6094L115.506 27.5508C115.201 27.4258 114.949 27.3385 114.748 27.2891C114.55 27.2396 114.355 27.2148 114.162 27.2148C113.933 27.2148 113.757 27.2682 113.635 27.375C113.513 27.4818 113.451 27.6211 113.451 27.793C113.451 27.8997 113.476 27.9935 113.526 28.0742C113.575 28.1523 113.653 28.2292 113.76 28.3047C113.869 28.3776 114.126 28.5104 114.53 28.7031C115.063 28.9583 115.429 29.2148 115.627 29.4727C115.825 29.7279 115.924 30.0417 115.924 30.4141Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
     </svg>
   );
