@@ -11,480 +11,480 @@ const ProphetPlot = () => {
     >
       <path
         d="M1.5 6C1.5 2.96244 3.96243 0.5 7 0.5H120C123.038 0.5 125.5 2.96243 125.5 6V119.796C125.5 122.834 123.038 125.296 120 125.296H7C3.96243 125.296 1.5 122.834 1.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M11.5 47.5L2 53V91.5L18 82L30 89.5L34.5 86L43.5 98L50 87.5L59 93.5L62 91.5L70.5 101.5L81.5 89.5L93 98L100.5 89.5L111 98L115 93.5L120 99.5L125 91.5V53L117.5 58.5L111 53L103 62.5L93 53L88 58.5L77.5 47.5L64.5 62.5L59 56.5L55.5 60L50 53L41.5 62.5L32.5 53L25.5 55L11.5 47.5Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <path
         d="M104.233 77.092C105.12 77.092 105.838 76.3678 105.838 75.4745C105.838 74.5811 105.12 73.8569 104.233 73.8569C103.347 73.8569 102.629 74.5811 102.629 75.4745C102.629 76.3678 103.347 77.092 104.233 77.092Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M106.729 78.8893C107.615 78.8893 108.333 78.1651 108.333 77.2717C108.333 76.3784 107.615 75.6542 106.729 75.6542C105.843 75.6542 105.125 76.3784 105.125 77.2717C105.125 78.1651 105.843 78.8893 106.729 78.8893Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.9"
         d="M113.403 74.3781C114.289 74.3781 115.007 73.654 115.007 72.7606C115.007 71.8673 114.289 71.1431 113.403 71.1431C112.517 71.1431 111.799 71.8673 111.799 72.7606C111.799 73.654 112.517 74.3781 113.403 74.3781Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.9"
         d="M106.016 72.7786C106.902 72.7786 107.62 72.0544 107.62 71.161C107.62 70.2677 106.902 69.5435 106.016 69.5435C105.13 69.5435 104.412 70.2677 104.412 71.161C104.412 72.0544 105.13 72.7786 106.016 72.7786Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.75"
         d="M104.947 70.2624C105.833 70.2624 106.551 69.5382 106.551 68.6449C106.551 67.7515 105.833 67.0273 104.947 67.0273C104.06 67.0273 103.342 67.7515 103.342 68.6449C103.342 69.5382 104.06 70.2624 104.947 70.2624Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.75"
         d="M106.016 68.4651C106.902 68.4651 107.62 67.7409 107.62 66.8476C107.62 65.9542 106.902 65.23 106.016 65.23C105.13 65.23 104.412 65.9542 104.412 66.8476C104.412 67.7409 105.13 68.4651 106.016 68.4651Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M107.799 67.0273C108.685 67.0273 109.403 66.3031 109.403 65.4098C109.403 64.5164 108.685 63.7922 107.799 63.7922C106.913 63.7922 106.194 64.5164 106.194 65.4098C106.194 66.3031 106.913 67.0273 107.799 67.0273Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M88.4474 78.6916C89.3334 78.6916 90.0517 77.9674 90.0517 77.0741C90.0517 76.1807 89.3334 75.4565 88.4474 75.4565C87.5614 75.4565 86.8431 76.1807 86.8431 77.0741C86.8431 77.9674 87.5614 78.6916 88.4474 78.6916Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M93.039 76.0556C93.925 76.0556 94.6433 75.3314 94.6433 74.4381C94.6433 73.5447 93.925 72.8205 93.039 72.8205C92.153 72.8205 91.4348 73.5447 91.4348 74.4381C91.4348 75.3314 92.153 76.0556 93.039 76.0556Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M94.6078 69.903C95.4938 69.903 96.2121 69.1788 96.2121 68.2854C96.2121 67.3921 95.4938 66.6679 94.6078 66.6679C93.7218 66.6679 93.0035 67.3921 93.0035 68.2854C93.0035 69.1788 93.7218 69.903 94.6078 69.903Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M97.8164 78.5298C98.7024 78.5298 99.4207 77.8056 99.4207 76.9123C99.4207 76.0189 98.7024 75.2947 97.8164 75.2947C96.9304 75.2947 96.2121 76.0189 96.2121 76.9123C96.2121 77.8056 96.9304 78.5298 97.8164 78.5298Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M99.2424 77.092C100.128 77.092 100.847 76.3678 100.847 75.4745C100.847 74.5811 100.128 73.8569 99.2424 73.8569C98.3564 73.8569 97.6381 74.5811 97.6381 75.4745C97.6381 76.3678 98.3564 77.092 99.2424 77.092Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M101.381 77.092C102.267 77.092 102.986 76.3678 102.986 75.4745C102.986 74.5811 102.267 73.8569 101.381 73.8569C100.495 73.8569 99.7772 74.5811 99.7772 75.4745C99.7772 76.3678 100.495 77.092 101.381 77.092Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M99.2424 67.0273C100.128 67.0273 100.847 66.3031 100.847 65.4098C100.847 64.5164 100.128 63.7922 99.2424 63.7922C98.3564 63.7922 97.6381 64.5164 97.6381 65.4098C97.6381 66.3031 98.3564 67.0273 99.2424 67.0273Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M80.6043 74.0187C81.4903 74.0187 82.2086 73.2945 82.2086 72.4011C82.2086 71.5078 81.4903 70.7836 80.6043 70.7836C79.7183 70.7836 79 71.5078 79 72.4011C79 73.2945 79.7183 74.0187 80.6043 74.0187Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M98.0731 82.6456C98.9591 82.6456 99.6774 81.9214 99.6774 81.028C99.6774 80.1347 98.9591 79.4105 98.0731 79.4105C97.1871 79.4105 96.4688 80.1347 96.4688 81.028C96.4688 81.9214 97.1871 82.6456 98.0731 82.6456Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M86.9811 70.2864C87.8671 70.2864 88.5854 69.5622 88.5854 68.6688C88.5854 67.7755 87.8671 67.0513 86.9811 67.0513C86.0951 67.0513 85.3768 67.7755 85.3768 68.6688C85.3768 69.5622 86.0951 70.2864 86.9811 70.2864Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M95.9086 68.6838C96.7946 68.6838 97.5129 67.9596 97.5129 67.0663C97.5129 66.1729 96.7946 65.4487 95.9086 65.4487C95.0226 65.4487 94.3043 66.1729 94.3043 67.0663C94.3043 67.9596 95.0226 68.6838 95.9086 68.6838Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M86.6649 68.6269C87.5509 68.6269 88.2692 67.9027 88.2692 67.0094C88.2692 66.116 87.5509 65.3918 86.6649 65.3918C85.7789 65.3918 85.0606 66.116 85.0606 67.0094C85.0606 67.9027 85.7789 68.6269 86.6649 68.6269Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M104.947 85C105.833 85 106.551 84.2758 106.551 83.3825C106.551 82.4891 105.833 81.7649 104.947 81.7649C104.06 81.7649 103.342 82.4891 103.342 83.3825C103.342 84.2758 104.06 85 104.947 85Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M116.761 76.0388C117.647 76.0388 118.365 75.3146 118.365 74.4213C118.365 73.5279 117.647 72.8037 116.761 72.8037C115.875 72.8037 115.157 73.5279 115.157 74.4213C115.157 75.3146 115.875 76.0388 116.761 76.0388Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M121.396 73.1632C122.282 73.1632 123 72.439 123 71.5456C123 70.6523 122.282 69.9281 121.396 69.9281C120.51 69.9281 119.791 70.6523 119.791 71.5456C119.791 72.439 120.51 73.1632 121.396 73.1632Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M109.884 69.3745C110.77 69.3745 111.488 68.6504 111.488 67.757C111.488 66.8637 110.77 66.1395 109.884 66.1395C108.998 66.1395 108.28 66.8637 108.28 67.757C108.28 68.6504 108.998 69.3745 109.884 69.3745Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M116.611 77.6132C117.497 77.6132 118.216 76.889 118.216 75.9957C118.216 75.1023 117.497 74.3781 116.611 74.3781C115.725 74.3781 115.007 75.1023 115.007 75.9957C115.007 76.889 115.725 77.6132 116.611 77.6132Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M108.815 74.7664C109.701 74.7664 110.419 74.0422 110.419 73.1488C110.419 72.2555 109.701 71.5313 108.815 71.5313C107.929 71.5313 107.21 72.2555 107.21 73.1488C107.21 74.0422 107.929 74.7664 108.815 74.7664Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M104.394 80.4996C105.28 80.4996 105.998 79.7754 105.998 78.8821C105.998 77.9888 105.28 77.2646 104.394 77.2646C103.508 77.2646 102.79 77.9888 102.79 78.8821C102.79 79.7754 103.508 80.4996 104.394 80.4996Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M105.82 81.9374C106.706 81.9374 107.424 81.2132 107.424 80.3199C107.424 79.4266 106.706 78.7024 105.82 78.7024C104.934 78.7024 104.216 79.4266 104.216 80.3199C104.216 81.2132 104.934 81.9374 105.82 81.9374Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M113.763 80.3379C114.649 80.3379 115.367 79.6137 115.367 78.7204C115.367 77.827 114.649 77.1028 113.763 77.1028C112.877 77.1028 112.159 77.827 112.159 78.7204C112.159 79.6137 112.877 80.3379 113.763 80.3379Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M115.189 78.9001C116.075 78.9001 116.793 78.1759 116.793 77.2825C116.793 76.3892 116.075 75.665 115.189 75.665C114.303 75.665 113.585 76.3892 113.585 77.2825C113.585 78.1759 114.303 78.9001 115.189 78.9001Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M117.328 78.9001C118.214 78.9001 118.932 78.1759 118.932 77.2825C118.932 76.3892 118.214 75.665 117.328 75.665C116.442 75.665 115.724 76.3892 115.724 77.2825C115.724 78.1759 116.442 78.9001 117.328 78.9001Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M114.02 84.4536C114.906 84.4536 115.624 83.7294 115.624 82.8361C115.624 81.9427 114.906 81.2185 114.02 81.2185C113.134 81.2185 112.415 81.9427 112.415 82.8361C112.415 83.7294 113.134 84.4536 114.02 84.4536Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M61.7665 71.908C60.8805 71.908 60.1622 72.6322 60.1622 73.5255C60.1622 74.4189 60.8805 75.1431 61.7665 75.1431C62.6525 75.1431 63.3708 74.4189 63.3708 73.5255C63.3708 72.6322 62.6525 71.908 61.7665 71.908Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M59.271 70.1107C58.3849 70.1107 57.6667 70.8349 57.6667 71.7283C57.6667 72.6216 58.3849 73.3458 59.271 73.3458C60.157 73.3458 60.8752 72.6216 60.8752 71.7283C60.8752 70.8349 60.157 70.1107 59.271 70.1107Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.9"
         d="M52.5971 74.6219C51.7111 74.6219 50.9929 75.346 50.9929 76.2394C50.9929 77.1327 51.7111 77.8569 52.5971 77.8569C53.4832 77.8569 54.2014 77.1327 54.2014 76.2394C54.2014 75.346 53.4832 74.6219 52.5971 74.6219Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.9"
         d="M59.984 76.2214C59.098 76.2214 58.3797 76.9456 58.3797 77.839C58.3797 78.7323 59.098 79.4565 59.984 79.4565C60.87 79.4565 61.5883 78.7323 61.5883 77.839C61.5883 76.9456 60.87 76.2214 59.984 76.2214Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.75"
         d="M61.0535 78.7376C60.1675 78.7376 59.4492 79.4618 59.4492 80.3551C59.4492 81.2485 60.1675 81.9727 61.0535 81.9727C61.9395 81.9727 62.6578 81.2485 62.6578 80.3551C62.6578 79.4618 61.9395 78.7376 61.0535 78.7376Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.75"
         d="M59.984 80.5349C59.098 80.5349 58.3797 81.2591 58.3797 82.1524C58.3797 83.0458 59.098 83.77 59.984 83.77C60.87 83.77 61.5883 83.0458 61.5883 82.1524C61.5883 81.2591 60.87 80.5349 59.984 80.5349Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M58.2014 81.9727C57.3154 81.9727 56.5972 82.6969 56.5972 83.5902C56.5972 84.4836 57.3154 85.2078 58.2014 85.2078C59.0875 85.2078 59.8057 84.4836 59.8057 83.5902C59.8057 82.6969 59.0875 81.9727 58.2014 81.9727Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M77.5526 70.3084C76.6666 70.3084 75.9483 71.0326 75.9483 71.9259C75.9483 72.8193 76.6666 73.5435 77.5526 73.5435C78.4386 73.5435 79.1569 72.8193 79.1569 71.9259C79.1569 71.0326 78.4386 70.3084 77.5526 70.3084Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M72.961 72.9444C72.075 72.9444 71.3567 73.6686 71.3567 74.5619C71.3567 75.4553 72.075 76.1795 72.961 76.1795C73.847 76.1795 74.5652 75.4553 74.5652 74.5619C74.5652 73.6686 73.847 72.9444 72.961 72.9444Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M71.3922 79.097C70.5062 79.097 69.7879 79.8212 69.7879 80.7146C69.7879 81.6079 70.5062 82.3321 71.3922 82.3321C72.2782 82.3321 72.9965 81.6079 72.9965 80.7146C72.9965 79.8212 72.2782 79.097 71.3922 79.097Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M68.1836 70.4702C67.2976 70.4702 66.5793 71.1944 66.5793 72.0877C66.5793 72.9811 67.2976 73.7053 68.1836 73.7053C69.0696 73.7053 69.7879 72.9811 69.7879 72.0877C69.7879 71.1944 69.0696 70.4702 68.1836 70.4702Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M66.7576 71.908C65.8716 71.908 65.1533 72.6322 65.1533 73.5255C65.1533 74.4189 65.8716 75.1431 66.7576 75.1431C67.6436 75.1431 68.3619 74.4189 68.3619 73.5255C68.3619 72.6322 67.6436 71.908 66.7576 71.908Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M64.6186 71.908C63.7325 71.908 63.0143 72.6322 63.0143 73.5255C63.0143 74.4189 63.7325 75.1431 64.6186 75.1431C65.5046 75.1431 66.2228 74.4189 66.2228 73.5255C66.2228 72.6322 65.5046 71.908 64.6186 71.908Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M66.7576 81.9727C65.8716 81.9727 65.1533 82.6969 65.1533 83.5902C65.1533 84.4836 65.8716 85.2078 66.7576 85.2078C67.6436 85.2078 68.3619 84.4836 68.3619 83.5902C68.3619 82.6969 67.6436 81.9727 66.7576 81.9727Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M85.3957 74.9813C84.5097 74.9813 83.7914 75.7055 83.7914 76.5989C83.7914 77.4922 84.5097 78.2164 85.3957 78.2164C86.2817 78.2164 87 77.4922 87 76.5989C87 75.7055 86.2817 74.9813 85.3957 74.9813Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M67.9269 66.3544C67.0409 66.3544 66.3226 67.0786 66.3226 67.972C66.3226 68.8653 67.0409 69.5895 67.9269 69.5895C68.8129 69.5895 69.5312 68.8653 69.5312 67.972C69.5312 67.0786 68.8129 66.3544 67.9269 66.3544Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M79.0189 78.7136C78.1329 78.7136 77.4146 79.4378 77.4146 80.3312C77.4146 81.2245 78.1329 81.9487 79.0189 81.9487C79.9049 81.9487 80.6232 81.2245 80.6232 80.3312C80.6232 79.4378 79.9049 78.7136 79.0189 78.7136Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M70.0914 80.3162C69.2054 80.3162 68.4871 81.0404 68.4871 81.9337C68.4871 82.8271 69.2054 83.5513 70.0914 83.5513C70.9774 83.5513 71.6957 82.8271 71.6957 81.9337C71.6957 81.0404 70.9774 80.3162 70.0914 80.3162Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M79.3351 80.3731C78.4491 80.3731 77.7308 81.0973 77.7308 81.9906C77.7308 82.884 78.4491 83.6082 79.3351 83.6082C80.2211 83.6082 80.9394 82.884 80.9394 81.9906C80.9394 81.0973 80.2211 80.3731 79.3351 80.3731Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M61.0535 64C60.1675 64 59.4492 64.7242 59.4492 65.6175C59.4492 66.5109 60.1675 67.2351 61.0535 67.2351C61.9395 67.2351 62.6578 66.5109 62.6578 65.6175C62.6578 64.7242 61.9395 64 61.0535 64Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M49.2389 72.9612C48.3528 72.9612 47.6346 73.6854 47.6346 74.5787C47.6346 75.4721 48.3528 76.1963 49.2389 76.1963C50.1249 76.1963 50.8431 75.4721 50.8431 74.5787C50.8431 73.6854 50.1249 72.9612 49.2389 72.9612Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M44.6043 75.8368C43.7183 75.8368 43 76.561 43 77.4544C43 78.3477 43.7183 79.0719 44.6043 79.0719C45.4903 79.0719 46.2086 78.3477 46.2086 77.4544C46.2086 76.561 45.4903 75.8368 44.6043 75.8368Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M56.1159 79.6255C55.2299 79.6255 54.5116 80.3496 54.5116 81.243C54.5116 82.1363 55.2299 82.8605 56.1159 82.8605C57.0019 82.8605 57.7202 82.1363 57.7202 81.243C57.7202 80.3496 57.0019 79.6255 56.1159 79.6255Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M49.3886 71.3868C48.5026 71.3868 47.7843 72.111 47.7843 73.0043C47.7843 73.8977 48.5026 74.6219 49.3886 74.6219C50.2746 74.6219 50.9929 73.8977 50.9929 73.0043C50.9929 72.111 50.2746 71.3868 49.3886 71.3868Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M57.1854 74.2336C56.2994 74.2336 55.5811 74.9578 55.5811 75.8512C55.5811 76.7445 56.2994 77.4687 57.1854 77.4687C58.0714 77.4687 58.7897 76.7445 58.7897 75.8512C58.7897 74.9578 58.0714 74.2336 57.1854 74.2336Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M61.606 68.5004C60.72 68.5004 60.0018 69.2246 60.0018 70.1179C60.0018 71.0112 60.72 71.7354 61.606 71.7354C62.4921 71.7354 63.2103 71.0112 63.2103 70.1179C63.2103 69.2246 62.4921 68.5004 61.606 68.5004Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M60.18 67.0626C59.294 67.0626 58.5757 67.7868 58.5757 68.6801C58.5757 69.5734 59.294 70.2976 60.18 70.2976C61.066 70.2976 61.7843 69.5734 61.7843 68.6801C61.7843 67.7868 61.066 67.0626 60.18 67.0626Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M52.2371 68.6621C51.3511 68.6621 50.6328 69.3863 50.6328 70.2796C50.6328 71.173 51.3511 71.8972 52.2371 71.8972C53.1231 71.8972 53.8414 71.173 53.8414 70.2796C53.8414 69.3863 53.1231 68.6621 52.2371 68.6621Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M50.8111 70.0999C49.925 70.0999 49.2068 70.8241 49.2068 71.7175C49.2068 72.6108 49.925 73.335 50.8111 73.335C51.6971 73.335 52.4153 72.6108 52.4153 71.7175C52.4153 70.8241 51.6971 70.0999 50.8111 70.0999Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M48.672 70.0999C47.786 70.0999 47.0677 70.8241 47.0677 71.7175C47.0677 72.6108 47.786 73.335 48.672 73.335C49.558 73.335 50.2763 72.6108 50.2763 71.7175C50.2763 70.8241 49.558 70.0999 48.672 70.0999Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M51.9804 64.5464C51.0944 64.5464 50.3761 65.2706 50.3761 66.1639C50.3761 67.0573 51.0944 67.7815 51.9804 67.7815C52.8664 67.7815 53.5847 67.0573 53.5847 66.1639C53.5847 65.2706 52.8664 64.5464 51.9804 64.5464Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M27.2335 76.092C28.1195 76.092 28.8378 75.3678 28.8378 74.4745C28.8378 73.5811 28.1195 72.8569 27.2335 72.8569C26.3475 72.8569 25.6292 73.5811 25.6292 74.4745C25.6292 75.3678 26.3475 76.092 27.2335 76.092Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M29.729 77.8893C30.6151 77.8893 31.3333 77.1651 31.3333 76.2717C31.3333 75.3784 30.6151 74.6542 29.729 74.6542C28.843 74.6542 28.1248 75.3784 28.1248 76.2717C28.1248 77.1651 28.843 77.8893 29.729 77.8893Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.9"
         d="M36.4029 73.3781C37.2889 73.3781 38.0071 72.654 38.0071 71.7606C38.0071 70.8673 37.2889 70.1431 36.4029 70.1431C35.5168 70.1431 34.7986 70.8673 34.7986 71.7606C34.7986 72.654 35.5168 73.3781 36.4029 73.3781Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.9"
         d="M29.016 71.7786C29.902 71.7786 30.6203 71.0544 30.6203 70.161C30.6203 69.2677 29.902 68.5435 29.016 68.5435C28.13 68.5435 27.4117 69.2677 27.4117 70.161C27.4117 71.0544 28.13 71.7786 29.016 71.7786Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.75"
         d="M27.9465 69.2624C28.8325 69.2624 29.5508 68.5382 29.5508 67.6449C29.5508 66.7515 28.8325 66.0273 27.9465 66.0273C27.0605 66.0273 26.3422 66.7515 26.3422 67.6449C26.3422 68.5382 27.0605 69.2624 27.9465 69.2624Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.75"
         d="M29.016 67.4651C29.902 67.4651 30.6203 66.7409 30.6203 65.8476C30.6203 64.9542 29.902 64.23 29.016 64.23C28.13 64.23 27.4117 64.9542 27.4117 65.8476C27.4117 66.7409 28.13 67.4651 29.016 67.4651Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M30.7986 66.0273C31.6846 66.0273 32.4028 65.3031 32.4028 64.4098C32.4028 63.5164 31.6846 62.7922 30.7986 62.7922C29.9125 62.7922 29.1943 63.5164 29.1943 64.4098C29.1943 65.3031 29.9125 66.0273 30.7986 66.0273Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M11.4474 77.6916C12.3334 77.6916 13.0517 76.9674 13.0517 76.0741C13.0517 75.1807 12.3334 74.4565 11.4474 74.4565C10.5614 74.4565 9.84314 75.1807 9.84314 76.0741C9.84314 76.9674 10.5614 77.6916 11.4474 77.6916Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M16.039 75.0556C16.925 75.0556 17.6433 74.3314 17.6433 73.4381C17.6433 72.5447 16.925 71.8205 16.039 71.8205C15.153 71.8205 14.4348 72.5447 14.4348 73.4381C14.4348 74.3314 15.153 75.0556 16.039 75.0556Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M17.6078 68.903C18.4938 68.903 19.2121 68.1788 19.2121 67.2854C19.2121 66.3921 18.4938 65.6679 17.6078 65.6679C16.7218 65.6679 16.0035 66.3921 16.0035 67.2854C16.0035 68.1788 16.7218 68.903 17.6078 68.903Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M20.8164 77.5298C21.7024 77.5298 22.4207 76.8056 22.4207 75.9123C22.4207 75.0189 21.7024 74.2947 20.8164 74.2947C19.9304 74.2947 19.2121 75.0189 19.2121 75.9123C19.2121 76.8056 19.9304 77.5298 20.8164 77.5298Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M22.2424 76.092C23.1284 76.092 23.8467 75.3678 23.8467 74.4745C23.8467 73.5811 23.1284 72.8569 22.2424 72.8569C21.3564 72.8569 20.6381 73.5811 20.6381 74.4745C20.6381 75.3678 21.3564 76.092 22.2424 76.092Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M24.3814 76.092C25.2675 76.092 25.9857 75.3678 25.9857 74.4745C25.9857 73.5811 25.2675 72.8569 24.3814 72.8569C23.4954 72.8569 22.7772 73.5811 22.7772 74.4745C22.7772 75.3678 23.4954 76.092 24.3814 76.092Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M22.2424 66.0273C23.1284 66.0273 23.8467 65.3031 23.8467 64.4098C23.8467 63.5164 23.1284 62.7922 22.2424 62.7922C21.3564 62.7922 20.6381 63.5164 20.6381 64.4098C20.6381 65.3031 21.3564 66.0273 22.2424 66.0273Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M3.60428 73.0187C4.4903 73.0187 5.20856 72.2945 5.20856 71.4011C5.20856 70.5078 4.4903 69.7836 3.60428 69.7836C2.71826 69.7836 2 70.5078 2 71.4011C2 72.2945 2.71826 73.0187 3.60428 73.0187Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M21.0731 81.6456C21.9591 81.6456 22.6774 80.9214 22.6774 80.028C22.6774 79.1347 21.9591 78.4105 21.0731 78.4105C20.1871 78.4105 19.4688 79.1347 19.4688 80.028C19.4688 80.9214 20.1871 81.6456 21.0731 81.6456Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M9.98111 69.2864C10.8671 69.2864 11.5854 68.5622 11.5854 67.6688C11.5854 66.7755 10.8671 66.0513 9.98111 66.0513C9.09509 66.0513 8.37683 66.7755 8.37683 67.6688C8.37683 68.5622 9.09509 69.2864 9.98111 69.2864Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M18.9086 67.6838C19.7946 67.6838 20.5129 66.9596 20.5129 66.0663C20.5129 65.1729 19.7946 64.4487 18.9086 64.4487C18.0226 64.4487 17.3043 65.1729 17.3043 66.0663C17.3043 66.9596 18.0226 67.6838 18.9086 67.6838Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M9.66489 67.6269C10.5509 67.6269 11.2692 66.9027 11.2692 66.0094C11.2692 65.116 10.5509 64.3918 9.66489 64.3918C8.77887 64.3918 8.06061 65.116 8.06061 66.0094C8.06061 66.9027 8.77887 67.6269 9.66489 67.6269Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M27.9465 84C28.8325 84 29.5508 83.2758 29.5508 82.3825C29.5508 81.4891 28.8325 80.7649 27.9465 80.7649C27.0605 80.7649 26.3422 81.4891 26.3422 82.3825C26.3422 83.2758 27.0605 84 27.9465 84Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M39.7611 75.0388C40.6472 75.0388 41.3654 74.3146 41.3654 73.4213C41.3654 72.5279 40.6472 71.8037 39.7611 71.8037C38.8751 71.8037 38.1569 72.5279 38.1569 73.4213C38.1569 74.3146 38.8751 75.0388 39.7611 75.0388Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M44.3957 72.1632C45.2817 72.1632 46 71.439 46 70.5456C46 69.6523 45.2817 68.9281 44.3957 68.9281C43.5097 68.9281 42.7914 69.6523 42.7914 70.5456C42.7914 71.439 43.5097 72.1632 44.3957 72.1632Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M32.8841 68.3745C33.7701 68.3745 34.4884 67.6504 34.4884 66.757C34.4884 65.8637 33.7701 65.1395 32.8841 65.1395C31.9981 65.1395 31.2798 65.8637 31.2798 66.757C31.2798 67.6504 31.9981 68.3745 32.8841 68.3745Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M39.6114 76.6132C40.4974 76.6132 41.2157 75.889 41.2157 74.9957C41.2157 74.1023 40.4974 73.3781 39.6114 73.3781C38.7254 73.3781 38.0071 74.1023 38.0071 74.9957C38.0071 75.889 38.7254 76.6132 39.6114 76.6132Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M31.8146 73.7664C32.7006 73.7664 33.4189 73.0422 33.4189 72.1488C33.4189 71.2555 32.7006 70.5313 31.8146 70.5313C30.9286 70.5313 30.2103 71.2555 30.2103 72.1488C30.2103 73.0422 30.9286 73.7664 31.8146 73.7664Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M27.394 79.4996C28.28 79.4996 28.9982 78.7754 28.9982 77.8821C28.9982 76.9888 28.28 76.2646 27.394 76.2646C26.5079 76.2646 25.7897 76.9888 25.7897 77.8821C25.7897 78.7754 26.5079 79.4996 27.394 79.4996Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M28.82 80.9374C29.706 80.9374 30.4243 80.2132 30.4243 79.3199C30.4243 78.4266 29.706 77.7024 28.82 77.7024C27.934 77.7024 27.2157 78.4266 27.2157 79.3199C27.2157 80.2132 27.934 80.9374 28.82 80.9374Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M36.7629 79.3379C37.6489 79.3379 38.3672 78.6137 38.3672 77.7204C38.3672 76.827 37.6489 76.1028 36.7629 76.1028C35.8769 76.1028 35.1586 76.827 35.1586 77.7204C35.1586 78.6137 35.8769 79.3379 36.7629 79.3379Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M38.1889 77.9001C39.075 77.9001 39.7932 77.1759 39.7932 76.2825C39.7932 75.3892 39.075 74.665 38.1889 74.665C37.3029 74.665 36.5847 75.3892 36.5847 76.2825C36.5847 77.1759 37.3029 77.9001 38.1889 77.9001Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M40.328 77.9001C41.214 77.9001 41.9323 77.1759 41.9323 76.2825C41.9323 75.3892 41.214 74.665 40.328 74.665C39.442 74.665 38.7237 75.3892 38.7237 76.2825C38.7237 77.1759 39.442 77.9001 40.328 77.9001Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         opacity="0.59"
         d="M37.0196 83.4536C37.9056 83.4536 38.6239 82.7294 38.6239 81.8361C38.6239 80.9427 37.9056 80.2185 37.0196 80.2185C36.1336 80.2185 35.4153 80.9427 35.4153 81.8361C35.4153 82.7294 36.1336 83.4536 37.0196 83.4536Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M18.1027 28.4648H18.7003C19.2589 28.4648 19.6769 28.3555 19.9542 28.1367C20.2316 27.9141 20.3702 27.5918 20.3702 27.1699C20.3702 26.7441 20.253 26.4297 20.0187 26.2266C19.7882 26.0234 19.4249 25.9219 18.9288 25.9219H18.1027V28.4648ZM22.2042 27.1055C22.2042 28.0273 21.9152 28.7324 21.337 29.2207C20.7628 29.709 19.9445 29.9531 18.882 29.9531H18.1027V33H16.2863V24.4336H19.0226C20.0616 24.4336 20.8507 24.6582 21.3898 25.1074C21.9327 25.5527 22.2042 26.2188 22.2042 27.1055ZM26.4778 28.2363H27.0638C27.638 28.2363 28.0618 28.1406 28.3352 27.9492C28.6087 27.7578 28.7454 27.457 28.7454 27.0469C28.7454 26.6406 28.6048 26.3516 28.3235 26.1797C28.0462 26.0078 27.6145 25.9219 27.0286 25.9219H26.4778V28.2363ZM26.4778 29.7129V33H24.6614V24.4336H27.1575C28.3216 24.4336 29.1829 24.6465 29.7415 25.0723C30.3001 25.4941 30.5794 26.1367 30.5794 27C30.5794 27.5039 30.4407 27.9531 30.1634 28.3477C29.886 28.7383 29.4934 29.0449 28.9856 29.2676C30.2747 31.1934 31.1145 32.4375 31.5052 33H29.4895L27.4446 29.7129H26.4778ZM41.1987 28.7051C41.1987 30.123 40.8471 31.2129 40.144 31.9746C39.4409 32.7363 38.433 33.1172 37.1205 33.1172C35.808 33.1172 34.8002 32.7363 34.0971 31.9746C33.394 31.2129 33.0424 30.1191 33.0424 28.6934C33.0424 27.2676 33.394 26.1797 34.0971 25.4297C34.8041 24.6758 35.8159 24.2988 37.1323 24.2988C38.4487 24.2988 39.4545 24.6777 40.1498 25.4355C40.8491 26.1934 41.1987 27.2832 41.1987 28.7051ZM34.9467 28.7051C34.9467 29.6621 35.1284 30.3828 35.4916 30.8672C35.8549 31.3516 36.3979 31.5938 37.1205 31.5938C38.5698 31.5938 39.2944 30.6309 39.2944 28.7051C39.2944 26.7754 38.5737 25.8105 37.1323 25.8105C36.4096 25.8105 35.8647 26.0547 35.4975 26.543C35.1303 27.0273 34.9467 27.748 34.9467 28.7051ZM45.6305 28.4648H46.2281C46.7867 28.4648 47.2047 28.3555 47.482 28.1367C47.7594 27.9141 47.898 27.5918 47.898 27.1699C47.898 26.7441 47.7809 26.4297 47.5465 26.2266C47.316 26.0234 46.9527 25.9219 46.4566 25.9219H45.6305V28.4648ZM49.732 27.1055C49.732 28.0273 49.443 28.7324 48.8648 29.2207C48.2906 29.709 47.4723 29.9531 46.4098 29.9531H45.6305V33H43.8141V24.4336H46.5504C47.5895 24.4336 48.3785 24.6582 48.9176 25.1074C49.4605 25.5527 49.732 26.2188 49.732 27.1055ZM59.2088 33H57.3982V29.3027H54.0056V33H52.1892V24.4336H54.0056V27.791H57.3982V24.4336H59.2088V33ZM67.1503 33H62.2167V24.4336H67.1503V25.9219H64.0331V27.8027H66.9335V29.291H64.0331V31.5H67.1503V33ZM73.088 33H71.2716V25.9453H68.9454V24.4336H75.4141V25.9453H73.088V33ZM83.3461 28.4648H83.9438C84.5023 28.4648 84.9203 28.3555 85.1977 28.1367C85.475 27.9141 85.6137 27.5918 85.6137 27.1699C85.6137 26.7441 85.4965 26.4297 85.2621 26.2266C85.0316 26.0234 84.6684 25.9219 84.1723 25.9219H83.3461V28.4648ZM87.4477 27.1055C87.4477 28.0273 87.1586 28.7324 86.5805 29.2207C86.0063 29.709 85.1879 29.9531 84.1254 29.9531H83.3461V33H81.5297V24.4336H84.266C85.3051 24.4336 86.0941 24.6582 86.6332 25.1074C87.1762 25.5527 87.4477 26.2188 87.4477 27.1055ZM89.9048 33V24.4336H91.7213V31.5H95.1959V33H89.9048ZM105.305 28.7051C105.305 30.123 104.954 31.2129 104.251 31.9746C103.548 32.7363 102.54 33.1172 101.227 33.1172C99.9148 33.1172 98.907 32.7363 98.2038 31.9746C97.5007 31.2129 97.1491 30.1191 97.1491 28.6934C97.1491 27.2676 97.5007 26.1797 98.2038 25.4297C98.9109 24.6758 99.9226 24.2988 101.239 24.2988C102.555 24.2988 103.561 24.6777 104.257 25.4355C104.956 26.1934 105.305 27.2832 105.305 28.7051ZM99.0534 28.7051C99.0534 29.6621 99.2351 30.3828 99.5984 30.8672C99.9616 31.3516 100.505 31.5938 101.227 31.5938C102.676 31.5938 103.401 30.6309 103.401 28.7051C103.401 26.7754 102.68 25.8105 101.239 25.8105C100.516 25.8105 99.9714 26.0547 99.6042 26.543C99.237 27.0273 99.0534 27.748 99.0534 28.7051ZM111.225 33H109.409V25.9453H107.083V24.4336H113.552V25.9453H111.225V33Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M1 74.4118L17.5 68.2353L36 79.3529L46.5 72.3529L52 76.0588L62.5 68.2353L76.5 81L85 70.2941L91.5 76.0588L99 70.2941L111 77.7059L125 67"
-        className="stroke-accent1"
+        className="stroke-accent2"
       />
     </svg>
   );
