@@ -11,174 +11,174 @@ const Scatter = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M36.7129 31.2558C36.7129 32.1582 36.3871 32.8691 35.7354 33.3887C35.0882 33.9082 34.1859 34.168 33.0283 34.168C31.9619 34.168 31.0186 33.9674 30.1982 33.5664V31.5976C30.8727 31.8984 31.4424 32.1103 31.9072 32.2334C32.3766 32.3564 32.805 32.418 33.1924 32.418C33.6572 32.418 34.0127 32.3291 34.2588 32.1514C34.5094 31.9736 34.6348 31.7093 34.6348 31.3584C34.6348 31.1624 34.5801 30.9892 34.4707 30.8389C34.3613 30.6839 34.1996 30.5358 33.9854 30.3945C33.7757 30.2532 33.3451 30.0277 32.6934 29.7178C32.0827 29.4306 31.6247 29.1549 31.3193 28.8906C31.014 28.6263 30.7702 28.3187 30.5879 27.9678C30.4056 27.6168 30.3145 27.2067 30.3145 26.7373C30.3145 25.8532 30.613 25.1582 31.21 24.6523C31.8115 24.1465 32.641 23.8935 33.6982 23.8935C34.2178 23.8935 34.7122 23.9551 35.1816 24.0781C35.6556 24.2012 36.1501 24.3743 36.665 24.5976L35.9815 26.2451C35.4482 26.0264 35.0062 25.8737 34.6553 25.7871C34.3089 25.7005 33.9671 25.6572 33.6299 25.6572C33.2288 25.6572 32.9212 25.7506 32.707 25.9375C32.4928 26.1243 32.3857 26.3681 32.3857 26.6689C32.3857 26.8558 32.429 27.0198 32.5156 27.1611C32.6022 27.2978 32.7389 27.4323 32.9258 27.5644C33.1172 27.692 33.5661 27.9245 34.2725 28.2617C35.2067 28.7083 35.847 29.1572 36.1934 29.6084C36.5397 30.055 36.7129 30.6042 36.7129 31.2558ZM43.7495 25.6572C42.952 25.6572 42.3345 25.958 41.897 26.5596C41.4595 27.1566 41.2407 27.9905 41.2407 29.0615C41.2407 31.29 42.077 32.4043 43.7495 32.4043C44.4514 32.4043 45.3013 32.2288 46.2993 31.8779V33.6553C45.479 33.9971 44.563 34.168 43.5513 34.168C42.0975 34.168 40.9855 33.7282 40.2154 32.8486C39.4452 31.9645 39.0601 30.6976 39.0601 29.0478C39.0601 28.0088 39.2492 27.0996 39.6275 26.3203C40.0057 25.5364 40.548 24.9372 41.2544 24.5224C41.9654 24.1032 42.7971 23.8935 43.7495 23.8935C44.7202 23.8935 45.6955 24.1282 46.6753 24.5976L45.9917 26.3203C45.618 26.1426 45.2421 25.9876 44.8638 25.8555C44.4855 25.7233 44.1141 25.6572 43.7495 25.6572ZM55.5303 34.0312L54.8057 31.6523H51.1622L50.4375 34.0312H48.1543L51.6817 23.9961H54.2725L57.8135 34.0312H55.5303ZM54.2999 29.875C53.6299 27.7194 53.2517 26.5003 53.1651 26.2178C53.0831 25.9352 53.0238 25.7119 52.9874 25.5478C52.837 26.1312 52.4063 27.5736 51.6954 29.875H54.2999ZM63.9136 34.0312H61.7945V25.8008H59.0806V24.0371H66.6275V25.8008H63.9136V34.0312ZM73.0011 34.0312H70.8819V25.8008H68.1681V24.0371H75.7149V25.8008H73.0011V34.0312ZM83.9889 34.0312H78.233V24.0371H83.9889V25.7734H80.3522V27.9678H83.7359V29.7041H80.3522V32.2812H83.9889V34.0312ZM89.1798 28.4736H89.8634C90.5333 28.4736 91.0278 28.362 91.3468 28.1387C91.6658 27.9153 91.8253 27.5644 91.8253 27.0859C91.8253 26.612 91.6613 26.2747 91.3331 26.0742C91.0096 25.8737 90.506 25.7734 89.8224 25.7734H89.1798V28.4736ZM89.1798 30.1963V34.0312H87.0607V24.0371H89.9728C91.3309 24.0371 92.3357 24.2855 92.9874 24.7822C93.6391 25.2744 93.965 26.0241 93.965 27.0312C93.965 27.6191 93.8032 28.1432 93.4796 28.6035C93.1561 29.0592 92.698 29.417 92.1056 29.6767C93.6095 31.9235 94.5893 33.375 95.045 34.0312H92.6935L90.3077 30.1963H89.1798Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <g clipPath="url(#clip0_1_79)">
         <path
           d="M69.898 89.8937C71.9309 89.8937 73.579 88.2458 73.579 86.2129C73.579 84.1801 71.9309 82.5321 69.898 82.5321C67.865 82.5321 66.217 84.1801 66.217 86.2129C66.217 88.2458 67.865 89.8937 69.898 89.8937Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M75.624 85.804C77.6569 85.804 79.305 84.156 79.305 82.1232C79.305 80.0903 77.6569 78.4424 75.624 78.4424C73.591 78.4424 71.943 80.0903 71.943 82.1232C71.943 84.156 73.591 85.804 75.624 85.804Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M75.624 93.9835C77.6569 93.9835 79.305 92.3355 79.305 90.3027C79.305 88.2699 77.6569 86.6219 75.624 86.6219C73.591 86.6219 71.943 88.2699 71.943 90.3027C71.943 92.3355 73.591 93.9835 75.624 93.9835Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           d="M78.896 93.9835C80.9289 93.9835 82.577 92.3355 82.577 90.3027C82.577 88.2699 80.9289 86.6219 78.896 86.6219C76.863 86.6219 75.215 88.2699 75.215 90.3027C75.215 92.3355 76.863 93.9835 78.896 93.9835Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.9"
           d="M90.9369 83.7182C92.9699 83.7182 94.6179 82.0702 94.6179 80.0374C94.6179 78.0045 92.9699 76.3566 90.9369 76.3566C88.904 76.3566 87.2559 78.0045 87.2559 80.0374C87.2559 82.0702 88.904 83.7182 90.9369 83.7182Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.9"
           d="M73.988 80.0783C76.0209 80.0783 77.669 78.4303 77.669 76.3975C77.669 74.3646 76.0209 72.7167 73.988 72.7167C71.955 72.7167 70.307 74.3646 70.307 76.3975C70.307 78.4303 71.955 80.0783 73.988 80.0783Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.75"
           d="M71.534 74.3526C73.5669 74.3526 75.215 72.7046 75.215 70.6718C75.215 68.639 73.5669 66.991 71.534 66.991C69.501 66.991 67.853 68.639 67.853 70.6718C67.853 72.7046 69.501 74.3526 71.534 74.3526Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.75"
           d="M73.988 70.2628C76.0209 70.2628 77.669 68.6149 77.669 66.582C77.669 64.5492 76.0209 62.9012 73.988 62.9012C71.955 62.9012 70.307 64.5492 70.307 66.582C70.307 68.6149 71.955 70.2628 73.988 70.2628Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M78.078 66.991C80.1109 66.991 81.759 65.3431 81.759 63.3102C81.759 61.2774 80.1109 59.6294 78.078 59.6294C76.045 59.6294 74.397 61.2774 74.397 63.3102C74.397 65.3431 76.045 66.991 78.078 66.991Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M33.677 93.5336C35.7099 93.5336 37.358 91.8857 37.358 89.8528C37.358 87.82 35.7099 86.172 33.677 86.172C31.644 86.172 29.996 87.82 29.996 89.8528C29.996 91.8857 31.644 93.5336 33.677 93.5336Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M44.2124 87.5353C46.2454 87.5353 47.8934 85.8873 47.8934 83.8545C47.8934 81.8217 46.2454 80.1737 44.2124 80.1737C42.1795 80.1737 40.5315 81.8217 40.5315 83.8545C40.5315 85.8873 42.1795 87.5353 44.2124 87.5353Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M47.812 73.5346C49.845 73.5346 51.493 71.8867 51.493 69.8538C51.493 67.821 49.845 66.173 47.812 66.173C45.779 66.173 44.131 67.821 44.131 69.8538C44.131 71.8867 45.779 73.5346 47.812 73.5346Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M55.174 93.1655C57.2069 93.1655 58.855 91.5176 58.855 89.4847C58.855 87.4519 57.2069 85.804 55.174 85.804C53.141 85.804 51.493 87.4519 51.493 89.4847C51.493 91.5176 53.141 93.1655 55.174 93.1655Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M58.446 89.8937C60.4789 89.8937 62.127 88.2458 62.127 86.2129C62.127 84.1801 60.4789 82.5321 58.446 82.5321C56.413 82.5321 54.765 84.1801 54.765 86.2129C54.765 88.2458 56.413 89.8937 58.446 89.8937Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M63.354 89.8937C65.3869 89.8937 67.035 88.2458 67.035 86.2129C67.035 84.1801 65.3869 82.5321 63.354 82.5321C61.321 82.5321 59.673 84.1801 59.673 86.2129C59.673 88.2458 61.321 89.8937 63.354 89.8937Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M58.446 66.991C60.4789 66.991 62.127 65.3431 62.127 63.3102C62.127 61.2774 60.4789 59.6294 58.446 59.6294C56.413 59.6294 54.765 61.2774 54.765 63.3102C54.765 65.3431 56.413 66.991 58.446 66.991Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M15.681 82.9002C17.714 82.9002 19.362 81.2523 19.362 79.2194C19.362 77.1866 17.714 75.5386 15.681 75.5386C13.648 75.5386 12 77.1866 12 79.2194C12 81.2523 13.648 82.9002 15.681 82.9002Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M55.763 102.531C57.7959 102.531 59.4439 100.883 59.4439 98.8503C59.4439 96.8175 57.7959 95.1695 55.763 95.1695C53.73 95.1695 52.082 96.8175 52.082 98.8503C52.082 100.883 53.73 102.531 55.763 102.531Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M30.3125 74.4071C32.3455 74.4071 33.9935 72.7592 33.9935 70.7263C33.9935 68.6935 32.3455 67.0455 30.3125 67.0455C28.2796 67.0455 26.6315 68.6935 26.6315 70.7263C26.6315 72.7592 28.2796 74.4071 30.3125 74.4071Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M50.7966 70.7604C52.8296 70.7604 54.4776 69.1125 54.4776 67.0796C54.4776 65.0468 52.8296 63.3988 50.7966 63.3988C48.7637 63.3988 47.1156 65.0468 47.1156 67.0796C47.1156 69.1125 48.7637 70.7604 50.7966 70.7604Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M29.587 70.6309C31.6199 70.6309 33.268 68.983 33.268 66.9501C33.268 64.9173 31.6199 63.2693 29.587 63.2693C27.554 63.2693 25.906 64.9173 25.906 66.9501C25.906 68.983 27.554 70.6309 29.587 70.6309Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M71.534 107.889C73.5669 107.889 75.215 106.241 75.215 104.208C75.215 102.175 73.5669 100.527 71.534 100.527C69.501 100.527 67.853 102.175 67.853 104.208C67.853 106.241 69.501 107.889 71.534 107.889Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M98.6425 87.4971C100.675 87.4971 102.323 85.8492 102.323 83.8163C102.323 81.7835 100.675 80.1355 98.6425 80.1355C96.6095 80.1355 94.9615 81.7835 94.9615 83.8163C94.9615 85.8492 96.6095 87.4971 98.6425 87.4971Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M109.276 80.9535C111.309 80.9535 112.957 79.3055 112.957 77.2727C112.957 75.2398 111.309 73.5919 109.276 73.5919C107.243 73.5919 105.595 75.2398 105.595 77.2727C105.595 79.3055 107.243 80.9535 109.276 80.9535Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M82.8633 72.3322C84.8962 72.3322 86.5443 70.6843 86.5443 68.6515C86.5443 66.6186 84.8962 64.9707 82.8633 64.9707C80.8303 64.9707 79.1823 66.6186 79.1823 68.6515C79.1823 70.6843 80.8303 72.3322 82.8633 72.3322Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M98.2989 91.0798C100.332 91.0798 101.98 89.4318 101.98 87.399C101.98 85.3661 100.332 83.7182 98.2989 83.7182C96.266 83.7182 94.6179 85.3661 94.6179 87.399C94.6179 89.4318 96.266 91.0798 98.2989 91.0798Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M80.4093 84.6016C82.4422 84.6016 84.0903 82.9536 84.0903 80.9208C84.0903 78.8879 82.4422 77.24 80.4093 77.24C78.3763 77.24 76.7283 78.8879 76.7283 80.9208C76.7283 82.9536 78.3763 84.6016 80.4093 84.6016Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M70.2661 97.6479C72.299 97.6479 73.9471 96 73.9471 93.9671C73.9471 91.9343 72.299 90.2863 70.2661 90.2863C68.2331 90.2863 66.5851 91.9343 66.5851 93.9671C66.5851 96 68.2331 97.6479 70.2661 97.6479Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M73.5381 100.92C75.571 100.92 77.2191 99.2718 77.2191 97.2389C77.2191 95.2061 75.571 93.5582 73.5381 93.5582C71.5051 93.5582 69.8571 95.2061 69.8571 97.2389C69.8571 99.2718 71.5051 100.92 73.5381 100.92Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M91.7631 97.2799C93.7961 97.2799 95.4441 95.6319 95.4441 93.5991C95.4441 91.5662 93.7961 89.9183 91.7631 89.9183C89.7301 89.9183 88.0821 91.5662 88.0821 93.5991C88.0821 95.6319 89.7301 97.2799 91.7631 97.2799Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M95.0351 94.008C97.068 94.008 98.7161 92.3601 98.7161 90.3272C98.7161 88.2944 97.068 86.6464 95.0351 86.6464C93.0021 86.6464 91.3541 88.2944 91.3541 90.3272C91.3541 92.3601 93.0021 94.008 95.0351 94.008Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M99.9431 94.008C101.976 94.008 103.624 92.3601 103.624 90.3272C103.624 88.2944 101.976 86.6464 99.9431 86.6464C97.9101 86.6464 96.2621 88.2944 96.2621 90.3272C96.2621 92.3601 97.9101 94.008 99.9431 94.008Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
         <path
           opacity="0.59"
           d="M92.3521 106.645C94.385 106.645 96.0331 104.997 96.0331 102.965C96.0331 100.932 94.385 99.2838 92.3521 99.2838C90.3191 99.2838 88.6711 100.932 88.6711 102.965C88.6711 104.997 90.3191 106.645 92.3521 106.645Z"
-          className="fill-accent1"
+          className="fill-accent2"
         />
       </g>
       <defs>
