@@ -11,20 +11,20 @@ const Table = () => {
     >
       <path
         d="M0.5 6C0.5 2.96244 2.96243 0.5 6 0.5H119C122.038 0.5 124.5 2.96243 124.5 6V119.796C124.5 122.834 122.038 125.296 119 125.296H6C2.96243 125.296 0.5 122.834 0.5 119.796V6Z"
-        className="fill-accent1 stroke-accent1"
+        className="fill-accent2 stroke-accent2"
         fillOpacity="0.05"
       />
       <path
         d="M43.9819 34.0312H41.8628V25.8008H39.1489V24.0371H46.6958V25.8008H43.9819V34.0312ZM55.3321 34.0312L54.6074 31.6523H50.9639L50.2393 34.0312H47.9561L51.4834 23.9961H54.0742L57.6153 34.0312H55.3321ZM54.1016 29.875C53.4317 27.7194 53.0534 26.5003 52.9668 26.2178C52.8848 25.9352 52.8255 25.7119 52.7891 25.5478C52.6387 26.1312 52.208 27.5736 51.4971 29.875H54.1016ZM59.8599 24.0371H62.9703C64.3876 24.0371 65.4152 24.2399 66.0533 24.6455C66.6958 25.0465 67.0171 25.6868 67.0171 26.5664C67.0171 27.1634 66.8759 27.6533 66.5933 28.0361C66.3153 28.4189 65.9439 28.6491 65.479 28.7265V28.7949C66.1125 28.9362 66.5682 29.2005 66.8462 29.5879C67.1288 29.9752 67.2701 30.4902 67.2701 31.1328C67.2701 32.0443 66.9397 32.7552 66.2788 33.2656C65.6226 33.776 64.7294 34.0312 63.5992 34.0312H59.8599V24.0371ZM61.979 27.9951H63.2095C63.7837 27.9951 64.1984 27.9062 64.4537 27.7285C64.7134 27.5508 64.8433 27.2568 64.8433 26.8467C64.8433 26.4639 64.702 26.1904 64.4195 26.0264C64.1415 25.8577 63.6994 25.7734 63.0933 25.7734H61.979V27.9951ZM61.979 29.6767V32.2812H63.3599C63.9432 32.2812 64.3739 32.1696 64.6519 31.9463C64.9299 31.723 65.0689 31.3812 65.0689 30.9209C65.0689 30.0915 64.4764 29.6767 63.2915 29.6767H61.979ZM70.2462 34.0312V24.0371H72.3653V32.2812H76.419V34.0312H70.2462ZM84.898 34.0312H79.1422V24.0371H84.898V25.7734H81.2613V27.9678H84.6451V29.7041H81.2613V32.2812H84.898V34.0312Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M26 50C26 48.8954 26.8954 48 28 48H98C99.1046 48 100 48.8954 100 50V57.9556H26V50Z"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <path
         d="M26 58H100V100C100 101.105 99.1046 102 98 102H28C26.8954 102 26 101.105 26 100V58Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -33,7 +33,7 @@ const Table = () => {
         width="20"
         height="8"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -42,7 +42,7 @@ const Table = () => {
         width="20"
         height="8"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
       />
       <rect
         x="48"
@@ -50,7 +50,7 @@ const Table = () => {
         width="20"
         height="8"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.8"
       />
       <rect
@@ -59,7 +59,7 @@ const Table = () => {
         width="20"
         height="8"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <rect
@@ -68,7 +68,7 @@ const Table = () => {
         width="20"
         height="8"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.6"
       />
       <rect
@@ -77,12 +77,12 @@ const Table = () => {
         width="20"
         height="8"
         rx="2"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.4"
       />
       <path
         d="M26 58H43V102H28C26.8954 102 26 101.105 26 100V58Z"
-        className="fill-accent1"
+        className="fill-accent2"
         fillOpacity="0.7"
       />
     </svg>
