@@ -1,7 +1,5 @@
-import { getBreakpointValue, useMantineTheme } from "@mantine/core";
 import { memo } from "react";
 import HeaderTab from "./HeaderTab";
-import { Logo } from "./Logo";
 import ControlBar from "../flow_chart_panel/views/ControlBar";
 import { useWindowSize } from "react-use";
 import { HEADER_HEIGHT } from "./Layout";
@@ -15,17 +13,17 @@ const tabs = [
   },
 ];
 
-const Header = () => {
-  const theme = useMantineTheme();
+const md = 800;
 
+const Header = () => {
   // Actual media query causes flickering... need to use this manual one
   const { width } = useWindowSize();
-  const large = width > getBreakpointValue(theme.breakpoints.sm);
+  const large = width > md;
 
   return (
-    <div style={{ height: HEADER_HEIGHT }}>
+    <div style={{ height: HEADER_HEIGHT }} className="">
       <div className="flex gap-4">
-        <Logo />
+        <img width={64} height={64} src="/assets/logo.png" alt="Logo" />
         {tabs.map((t) => (
           <HeaderTab to={t.to} testId={t.testId} key={t.fullText}>
             {large ? t.fullText : t.shortText}
