@@ -5,9 +5,9 @@ type LayoutProps = {
   children: React.ReactNode;
 };
 
-export const HEADER_HEIGHT = 70;
-export const ACTIONS_HEIGHT = 57;
-const SERVER_STATUS_HEIGHT = 30;
+export const HEADER_HEIGHT = 72;
+export const ACTIONS_HEIGHT = 56;
+const SERVER_STATUS_HEIGHT = 32;
 
 export const LAYOUT_TOP_HEIGHT =
   HEADER_HEIGHT + ACTIONS_HEIGHT + SERVER_STATUS_HEIGHT;
@@ -19,7 +19,7 @@ export const Layout = ({ children }: LayoutProps) => {
 
   return (
     <div>
-      <div className="sm:px-8">
+      <div className="mx-8">
         <div
           data-cy="app-status"
           id="app-status"
