@@ -1,4 +1,3 @@
-import "flojoy/styles/styles.css";
 import { useFlowChartState } from "@hooks/useFlowChartState";
 import { useMantineTheme } from "@mantine/core";
 import PYTHON_FUNCTIONS from "@src/data/pythonFunctions.json";
@@ -33,18 +32,19 @@ import { useFlowChartTabState } from "./FlowChartTabState";
 import { useAddNewNode } from "./hooks/useAddNewNode";
 import { NodeExpandMenu } from "./views/NodeExpandMenu";
 import { sendEventToMix } from "@src/services/MixpanelServices";
-import { ACTIONS_HEIGHT, Layout } from "../common/Layout";
+import { ACTIONS_HEIGHT, LAYOUT_TOP_HEIGHT, Layout } from "../common/Layout";
 import { getEdgeTypes, isCompatibleType } from "@src/utils/TypeCheck";
 import { CenterObserver } from "./components/CenterObserver";
-import { CommandMenu } from "../command/CommandMenu";
+// import { CommandMenu } from "../command/CommandMenu";
 import useNodeTypes from "./hooks/useNodeTypes";
 import { Separator } from "@src/components/ui/separator";
-import { Workflow } from "lucide-react";
-import { IconButton } from "../common/IconButton";
+import { Pencil, Workflow, X } from "lucide-react";
 import { GalleryModal } from "@src/components/gallery/GalleryModal";
 import { toast, Toaster } from "sonner";
-import { useTheme } from "@src/components/theme-provider";
+import { useTheme } from "@src/providers/theme-provider";
 import { ClearCanvasBtn } from "./components/ClearCanvasBtn";
+import { Button } from "@src/components/ui/button";
+import { ResizeFitter } from "./components/ResizeFitter";
 
 localforage.config({
   name: "react-flow",
@@ -57,8 +57,13 @@ const FlowChartTab = () => {
 
   const { theme } = useTheme();
 
-  const { isSidebarOpen, setIsSidebarOpen, setRfInstance, setIsEditMode } =
-    useFlowChartState();
+  const {
+    isSidebarOpen,
+    setIsSidebarOpen,
+    setRfInstance,
+    isEditMode,
+    setIsEditMode,
+  } = useFlowChartState();
 
   const mantineTheme = useMantineTheme();
 
@@ -66,16 +71,8 @@ const FlowChartTab = () => {
     states: { programResults },
   } = useSocket();
 
-  const {
-    nodeLabel,
-    nodeType,
-    pythonString,
-    setPythonString,
-    nodeFilePath,
-    setNodeFilePath,
-    setNodeLabel,
-    setNodeType,
-  } = useFlowChartTabState();
+  const { pythonString, setPythonString, nodeFilePath, setNodeFilePath } =
+    useFlowChartTabState();
 
   const { nodes, setNodes, edges, setEdges, selectedNode, unSelectedNodes } =
     useFlowChartGraph();
@@ -114,17 +111,19 @@ const FlowChartTab = () => {
     () => ({ default: SmartBezierEdge }),
     [],
   );
-  // Attach a callback to each of the custom nodes.
-  // This is to pass down the setNodes/setEdges functions as props for deleting nodes.
-  // Has to be passed through the data prop because passing as a regular prop doesn't work
-  // for whatever reason.
+
   const nodeTypes = useNodeTypes({
     handleRemove: handleNodeRemove,
-    wrapperOnClick: () => setIsEditMode(true),
+    wrapperOnClick: () => {
+      setIsEditMode(true);
+    },
     theme: mantineTheme.colorScheme,
   });
+
   const onInit: OnInit = (rfIns) => {
-    rfIns.fitView();
+    rfIns.fitView({
+      padding: 0.8,
+    });
     setRfInstance(rfIns.toObject());
   };
   const handleNodeDrag: NodeDragHandler = (_, node) => {
@@ -183,15 +182,7 @@ const FlowChartTab = () => {
     const nodeFileData = PYTHON_FUNCTIONS[nodeFileName] ?? {};
     setNodeFilePath(nodeFileData.path ?? "");
     setPythonString(nodeFileData.metadata ?? "");
-    setNodeLabel(selectedNode.data.label);
-    setNodeType(selectedNode.data.type);
-  }, [
-    selectedNode,
-    setNodeFilePath,
-    setNodeLabel,
-    setNodeType,
-    setPythonString,
-  ]);
+  }, [selectedNode, setNodeFilePath, setPythonString]);
 
   const proOptions = { hideAttribution: true };
 
@@ -232,40 +223,65 @@ const FlowChartTab = () => {
 
   return (
     <Layout>
-      <div className="sm:px-8" style={{ height: ACTIONS_HEIGHT }}>
-        <div className="py-1" />
-        <div className="flex">
-          <IconButton
-            icon={Workflow}
-            onClick={toggleSidebar}
-            data-testid="add-node-button"
-            variant="ghost"
-          >
-            Add Node
-          </IconButton>
-          <GalleryModal
-            isGalleryOpen={isGalleryOpen}
-            setIsGalleryOpen={setIsGalleryOpen}
-          />
-          <div className="grow" />
-          <ClearCanvasBtn clearCanvas={clearCanvas} />
-        </div>
-        <div className="py-1" />
-        <Separator />
-      </div>
-
-      <Sidebar
-        sections={nodeSection}
-        leafNodeClickHandler={addNewNode as LeafClickHandler}
-        isSideBarOpen={isSidebarOpen}
-        setSideBarStatus={setIsSidebarOpen}
-      />
-
-      <Toaster theme={theme} />
-
       <ReactFlowProvider>
+        <div className="mx-8" style={{ height: ACTIONS_HEIGHT }}>
+          <div className="py-1" />
+          <div className="flex">
+            <Button
+              data-testid="add-node-button"
+              className="gap-2"
+              variant="ghost"
+              onClick={toggleSidebar}
+            >
+              <Workflow size={20} className="stroke-muted-foreground" />
+              Add Node
+            </Button>
+
+            <GalleryModal
+              isGalleryOpen={isGalleryOpen}
+              setIsGalleryOpen={setIsGalleryOpen}
+            />
+            <div className="grow" />
+            {selectedNode && (
+              <>
+                {!isEditMode ? (
+                  <Button
+                    variant="ghost"
+                    className="gap-2"
+                    onClick={() => setIsEditMode(true)}
+                  >
+                    <Pencil size={18} className="stroke-muted-foreground" />
+                    Edit Node
+                  </Button>
+                ) : (
+                  <Button
+                    variant="ghost"
+                    className="gap-2"
+                    onClick={() => setIsEditMode(false)}
+                  >
+                    <X size={18} className="stroke-muted-foreground" />
+                    Cancel Edit
+                  </Button>
+                )}
+              </>
+            )}
+            <ClearCanvasBtn clearCanvas={clearCanvas} />
+          </div>
+          <div className="py-1" />
+          <Separator />
+        </div>
+
+        <Sidebar
+          sections={nodeSection}
+          leafNodeClickHandler={addNewNode as LeafClickHandler}
+          isSideBarOpen={isSidebarOpen}
+          setSideBarStatus={setIsSidebarOpen}
+        />
+
+        <Toaster theme={theme} />
+
         <div
-          style={{ height: "calc(100vh - 150px)" }}
+          style={{ height: `calc(100vh - ${LAYOUT_TOP_HEIGHT}px)` }}
           data-testid="react-flow"
           data-rfinstance={JSON.stringify(nodes)}
         >
@@ -274,13 +290,12 @@ const FlowChartTab = () => {
               nodes.filter((n) => n.selected).length > 1 ? null : selectedNode
             }
             unSelectedNodes={unSelectedNodes}
-            nodes={nodes}
-            setNodes={setNodes}
             setNodeModalOpen={() => setNodeModalOpen(true)}
             handleDelete={handleNodeRemove}
           />
 
           <FlowChartKeyboardShortcuts />
+          <ResizeFitter />
           <CenterObserver />
 
           <ReactFlow
@@ -303,7 +318,9 @@ const FlowChartTab = () => {
             onConnect={onConnect}
             onNodeDragStop={handleNodeDrag}
             onNodesDelete={handleNodesDelete}
-            fitView
+            fitViewOptions={{
+              padding: 0.8,
+            }}
           >
             <MiniMap
               style={{
@@ -325,22 +342,20 @@ const FlowChartTab = () => {
               zoomable
               pannable
             />
-            <Controls />
+            <Controls fitViewOptions={{ padding: 0.8 }} />
           </ReactFlow>
 
           <NodeExpandMenu
             selectedNode={selectedNode}
-            closeModal={() => setNodeModalOpen(false)}
             modalIsOpen={nodeModalOpen}
+            setModalOpen={setNodeModalOpen}
             nodeResults={programResults}
-            nodeLabel={nodeLabel}
-            nodeType={nodeType}
             pythonString={pythonString}
             nodeFilePath={nodeFilePath}
           />
         </div>
       </ReactFlowProvider>
-      <CommandMenu />
+      {/* <CommandMenu /> */}
     </Layout>
   );
 };
