@@ -1,21 +1,18 @@
 import { useSocket } from "@src/hooks/useSocket";
-import { nodeTypesMap } from "flojoy/components";
-import { useMemo } from "react";
+import { nodeTypesMap } from "@/components/nodes/nodeTypesMap";
+import React, { useMemo } from "react";
 import { NodeTypes } from "reactflow";
+import { MouseEvent } from "react";
 
 type UseNodeTypesProps = {
   handleRemove: (nodeId: string, nodeLabel: string) => void;
-  wrapperOnClick: () => void;
+  wrapperOnClick: (event: MouseEvent<HTMLDivElement>) => void;
   theme: "light" | "dark";
 };
 
-const useNodeTypes = ({
-  handleRemove,
-  wrapperOnClick,
-  theme,
-}: UseNodeTypesProps) => {
+const useNodeTypes = ({ handleRemove, theme }: UseNodeTypesProps) => {
   const {
-    states: { programResults, failedNode, runningNode, failureReason },
+    states: { programResults, failedNodes, runningNode },
   } = useSocket();
 
   const nodeTypes: NodeTypes = useMemo(
@@ -31,14 +28,11 @@ const useNodeTypes = ({
               return (
                 <CustomNode
                   isRunning={runningNode === props.data.id}
-                  nodeError={
-                    failedNode === props.id ? failureReason : undefined
-                  }
+                  nodeError={failedNodes[props.id]}
                   plotlyFig={nodeResult?.result.plotly_fig ?? undefined}
                   textBlob={nodeResult?.result.text_blob ?? undefined}
                   nodeProps={props}
                   handleRemove={handleRemove}
-                  wrapperOnClick={wrapperOnClick}
                   theme={theme}
                 />
               );
@@ -50,7 +44,7 @@ const useNodeTypes = ({
     // infinite re-render, so exception for eslint eslint-disable-next-line react-hooks/exhaustive-deps is added
     // to suppress eslint warning
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [programResults, failedNode, runningNode, failureReason, theme],
+    [programResults, failedNodes, runningNode, theme],
   );
 
   return nodeTypes;
