@@ -1,7 +1,7 @@
 const { createProxyMiddleware } = require("http-proxy-middleware");
 
 const BACKEND_HOST = process.env.REACT_APP_BACKEND_HOST || "127.0.0.1";
-const BACKEND_PORT = +process.env.REACT_APP_BACKEND_PORT || 8000;
+const BACKEND_PORT = +process.env.REACT_APP_BACKEND_PORT || 5392;
 const target = "http://" + BACKEND_HOST + ":" + BACKEND_PORT;
 
 module.exports = function (app) {
